Reorganise old 'types' into 'examples'
