+++ app/main.js
@@ -16,7 +16,7 @@
   \************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ AvO)\n/* harmony export */ });\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics.js */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_story_types_example_story_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/story/types/example-story.js */ \"./src/avo/story/types/example-story.js\");\n/* harmony import */ var _avo_interaction__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/interaction */ \"./src/avo/interaction/index.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\n\n\n // TODO: remove, or replace with a simpler version\n\nvar searchParams = new URLSearchParams(window.location.search);\nvar DEBUG = searchParams.get('debug') || false;\nvar AvO = /*#__PURE__*/function () {\n  function AvO() {\n    var _this$story;\n    var args = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n    _classCallCheck(this, AvO);\n    var _args$story = args.story,\n      story = _args$story === void 0 ? _avo_story_types_example_story_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"] : _args$story,\n      _args$width = args.width,\n      width = _args$width === void 0 ? 24 * _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$width,\n      _args$height = args.height,\n      height = _args$height === void 0 ? 16 * _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$height;\n    this.html = {\n      main: document.getElementById('main'),\n      canvas: document.getElementById('canvas'),\n      homeMenu: document.getElementById('home-menu'),\n      interactionMenu: document.getElementById('interaction-menu'),\n      buttonHome: document.getElementById('button-home'),\n      buttonFullscreen: document.getElementById('button-fullscreen'),\n      buttonReload: document.getElementById('button-reload')\n    };\n    this.homeMenu = false;\n    this.setHomeMenu(false);\n    this.interactionMenu = false;\n    this.setInteractionMenu(false);\n    this.canvas2d = this.html.canvas.getContext('2d');\n    this.canvasWidth = width;\n    this.canvasHeight = height;\n    this._canvasHasCameraTransforms = false; // Safety check\n\n    this.camera = {\n      target: null,\n      // Target entity to follow. If null, camera is static.\n      x: 0,\n      y: 0,\n      zoom: 1\n    };\n    this.setupUI();\n    this.entities = []; // Game objects\n    this.hero = null; // Main player-controlled entity\n    this.gameMap = {\n      // Game map and environment\n      tiles: [],\n      // 2D array of map tiles\n      width: 0,\n      height: 0\n    };\n    this.rules = new Map(); // Game rules, including win/loss conditions and player controls\n\n    this.story = story ? new story(this) : undefined; // Game story, i.e. the \"container\" that describes an adventure or experience, which contains game objects and game rules and etc\n    this.assets = ((_this$story = this.story) === null || _this$story === void 0 ? void 0 : _this$story.assets) || {};\n    this.secretAssets = {};\n    this.eventListeners = {\n      'keydown': [],\n      'keyup': [],\n      'pointerdown': [],\n      'pointermove': [],\n      'pointerup': [],\n      'pointertap': [],\n      'pointerholdend': []\n    };\n    this.playerInput = {};\n    this.resetPlayerInput();\n    this.initialised = false;\n    this.timeAccumulator = 0;\n    this.prevTime = null;\n    this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n  }\n  _createClass(AvO, [{\n    key: \"initialisationCheck\",\n    value: function initialisationCheck() {\n      var _this = this;\n      // Assets check\n      var allAssetsReady = true;\n      var numReadyAssets = 0;\n      var numTotalAssets = 0;\n      Object.keys(this.assets).forEach(function (id) {\n        var asset = _this.assets[id];\n        allAssetsReady = allAssetsReady && asset.ready;\n        if (asset.ready) numReadyAssets++;\n        numTotalAssets++;\n      });\n      Object.keys(this.secretAssets).forEach(function (id) {\n        var secretAsset = _this.secretAssets[id];\n        var secretAssetIsReady = secretAsset.ready || secretAsset.error;\n        allAssetsReady = allAssetsReady && secretAssetIsReady;\n        if (secretAssetIsReady) numReadyAssets++;\n        numTotalAssets++;\n      });\n\n      // Paint status\n      this.canvas2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      this.canvas2d.textAlign = 'start';\n      this.canvas2d.textBaseline = 'top';\n      this.canvas2d.fillStyle = '#ccc';\n      this.canvas2d.font = \"1em monospace\";\n      this.canvas2d.fillText(\"Loading \".concat(numReadyAssets, \" / \").concat(numTotalAssets, \" \"), _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n      if (allAssetsReady) {\n        var _this$story2;\n        // Clean up secret assets\n        Object.keys(this.secretAssets).forEach(function (id) {\n          if (_this.secretAssets[id].error) delete _this.secretAssets[id];\n        });\n\n        // Let's go!\n        this.initialised = true;\n        this.showUI();\n        (_this$story2 = this.story) === null || _this$story2 === void 0 || _this$story2.start();\n      }\n    }\n\n    /*\r\n    Section: General Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    The main loop. Run a single frame of gameplay.\r\n    - time: the current/total time (milliseconds) since the game started.\r\n     */\n  }, {\n    key: \"main\",\n    value: function main(time) {\n      var timeStep = this.prevTime ? time - this.prevTime : time;\n      this.prevTime = time;\n      this.timeAccumulator += timeStep;\n\n      // Sanity/safety check: timeStep can be huge\n      // e.g. if player pauses game by switching windows.\n      this.timeAccumulator = Math.min(this.timeAccumulator, _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION * 10);\n      if (this.initialised) {\n        // Keep a consistent \"frame rate\" for logic processing\n        while (this.timeAccumulator >= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION) {\n          this.play();\n          this.timeAccumulator -= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION;\n          // TODO: add safety counter to prevent excessively long while() loops.\n        }\n        // Paint whenever possible\n        this.paint();\n      } else {\n        this.initialisationCheck();\n      }\n      this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n    }\n\n    /*\r\n    Run the gameplay/physics logic for a single frame.\r\n     */\n  }, {\n    key: \"play\",\n    value: function play() {\n      var _this2 = this;\n      // If a menu is open, pause all action gameplay\n      if (this.homeMenu || this.interactionMenu) return;\n\n      // Run the action gameplay\n      // ----------------\n      this.rules.forEach(function (rule) {\n        return rule.play();\n      });\n      this.entities.forEach(function (entity) {\n        return entity.play();\n      });\n      this.checkCollisions();\n\n      // Cleanup: entities\n      this.entities.filter(function (entity) {\n        return entity._expired;\n      }).forEach(function (entity) {\n        return entity.deconstructor();\n      });\n      this.entities = this.entities.filter(function (entity) {\n        return !entity._expired;\n      });\n\n      // Cleanup: rules\n      this.rules.forEach(function (rule, id) {\n        if (rule.expired) {\n          rule.deconstructor();\n          _this2.rules[\"delete\"](id);\n        }\n      });\n\n      // Sort Entities along the y-axis, for paint()/rendering purposes.\n      // WARNING: inefficient\n      this.entities.sort(function (a, b) {\n        return a.y - b.y;\n      });\n      // ----------------\n\n      // Increment the duration of each currently pressed key\n      Object.keys(this.playerInput.keysPressed).forEach(function (key) {\n        if (_this2.playerInput.keysPressed[key]) _this2.playerInput.keysPressed[key].duration += _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION;\n      });\n\n      // Increment the duration of the pointer being active\n      if (this.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN) {\n        this.playerInput.pointerDownDuration += _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION;\n      }\n    }\n\n    /*\r\n    Paint/draw the game visuals onto the canvas.\r\n     */\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var _this3 = this;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n\n      // Camera Controls: focus the camera on the target entity, if any.\n      // ----------------\n      if (camera.target) {\n        camera.x = this.canvasWidth / 2 - camera.target.x * camera.zoom;\n        camera.y = this.canvasHeight / 2 - camera.target.y * camera.zoom;\n      }\n      c2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      c2d.resetTransform();\n      c2d.strokeStyle = 'rgba(128, 128, 128, 0.05)';\n      c2d.lineWidth = 2;\n      // ----------------\n\n      // Draw background grid\n      // ----------------\n      var GRID_SIZE = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * camera.zoom;\n      var offsetX = this.camera.x % GRID_SIZE - GRID_SIZE;\n      var offsetY = this.camera.y % GRID_SIZE - GRID_SIZE;\n      for (var y = offsetY; y < this.canvasHeight; y += GRID_SIZE) {\n        for (var x = offsetX; x < this.canvasWidth; x += GRID_SIZE) {\n          c2d.beginPath();\n          c2d.rect(x, y, GRID_SIZE, GRID_SIZE);\n          c2d.stroke();\n\n          // Debug Grid\n          if (DEBUG) {\n            c2d.fillStyle = '#ccc';\n            c2d.font = \"\".concat(camera.zoom * 0.5, \"em Source Code Pro\");\n            c2d.textAlign = 'center';\n            c2d.textBaseline = 'middle';\n            var col = Math.floor((x - this.camera.x) / GRID_SIZE);\n            var row = Math.floor((y - this.camera.y) / GRID_SIZE);\n            c2d.fillText(col + ',' + row, x + GRID_SIZE / 2, y + GRID_SIZE / 2); // using template strings here messes up colours in Brackets.\n          }\n        }\n      }\n      // ----------------\n\n      // Draw map tiles, entities, and other elements\n      // ----------------\n      var _loop = function _loop(layer) {\n        // 1. draw map tiles first, as these form the background. \n        for (var _row = 0; _row < _this3.gameMap.height; _row++) {\n          for (var _col = 0; _col < _this3.gameMap.width; _col++) {\n            var _this3$gameMap$tiles;\n            (_this3$gameMap$tiles = _this3.gameMap.tiles) === null || _this3$gameMap$tiles === void 0 || (_this3$gameMap$tiles = _this3$gameMap$tiles[_row]) === null || _this3$gameMap$tiles === void 0 || (_this3$gameMap$tiles = _this3$gameMap$tiles[_col]) === null || _this3$gameMap$tiles === void 0 || _this3$gameMap$tiles.paint(layer);\n          }\n        }\n\n        // 2. draw entities next.\n        _this3.entities.forEach(function (entity) {\n          return entity.paint(layer);\n        });\n\n        // 3. draw rules last, as these usually represent HUD visualisations.\n        _this3.rules.forEach(function (rule) {\n          return rule.paint(layer);\n        });\n      };\n      for (var layer = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.MIN_LAYER; layer <= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.MAX_LAYER; layer++) {\n        _loop(layer);\n      }\n      // ----------------\n    }\n\n    /*\r\n    Section: UI\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"setupUI\",\n    value: function setupUI() {\n      this.html.canvas.width = this.canvasWidth;\n      this.html.canvas.height = this.canvasHeight;\n      this.canvas2d.imageSmoothingEnabled = false; /* Pixel art: Maintains sprites' pixel sharpness when scaled up via drawImage() */\n\n      if (window.PointerEvent) {\n        this.html.canvas.addEventListener('pointerdown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('pointermove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('pointerup', this.onPointerUp.bind(this));\n        this.html.canvas.addEventListener('pointercancel', this.onPointerUp.bind(this));\n      } else {\n        this.html.canvas.addEventListener('mousedown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('mousemove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('mouseup', this.onPointerUp.bind(this));\n      }\n\n      // Prevent \"touch and hold to open context menu\" menu on touchscreens.\n      this.html.canvas.addEventListener('touchstart', stopEvent);\n      this.html.canvas.addEventListener('touchmove', stopEvent);\n      this.html.canvas.addEventListener('touchend', stopEvent);\n      this.html.canvas.addEventListener('touchcancel', stopEvent);\n      this.html.buttonHome.addEventListener('click', this.buttonHome_onClick.bind(this));\n      this.html.buttonFullscreen.addEventListener('click', this.buttonFullscreen_onClick.bind(this));\n      this.html.buttonReload.addEventListener('click', this.buttonReload_onClick.bind(this));\n      this.html.main.addEventListener('keydown', this.onKeyDown.bind(this));\n      this.html.main.addEventListener('keyup', this.onKeyUp.bind(this));\n      window.addEventListener('resize', this.updateUI.bind(this));\n      this.updateUI();\n      this.hideUI(); // Hide until all assets are ready\n\n      this.html.main.focus();\n    }\n  }, {\n    key: \"hideUI\",\n    value: function hideUI() {\n      this.html.buttonHome.style.visibility = 'hidden';\n      this.html.buttonReload.style.visibility = 'hidden';\n    }\n  }, {\n    key: \"showUI\",\n    value: function showUI() {\n      this.html.buttonHome.style.visibility = 'visible';\n      this.html.buttonReload.style.visibility = 'visible';\n    }\n  }, {\n    key: \"updateUI\",\n    value: function updateUI() {\n      // Fit the interaction layers (menus, etc) to the canvas\n      var mainDivBounds = this.html.main.getBoundingClientRect();\n      var canvasBounds = this.html.canvas.getBoundingClientRect();\n      this.html.homeMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.homeMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.homeMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.homeMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n      this.html.interactionMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.interactionMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.interactionMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.interactionMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n    }\n  }, {\n    key: \"setHomeMenu\",\n    value: function setHomeMenu(homeMenu) {\n      this.homeMenu = homeMenu;\n      if (homeMenu) {\n        this.html.homeMenu.style.visibility = 'visible';\n        this.html.buttonReload.style.visibility = 'hidden';\n      } else {\n        this.html.homeMenu.style.visibility = 'hidden';\n        this.html.buttonReload.style.visibility = 'visible';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"setInteractionMenu\",\n    value: function setInteractionMenu(interactionMenu) {\n      var div = this.html.interactionMenu;\n      this.interactionMenu && this.interactionMenu.unload(); // Unload the old menu, if any\n      this.interactionMenu = interactionMenu; // Set the new menu\n\n      if (interactionMenu) {\n        while (div.firstChild) {\n          div.removeChild(div.firstChild);\n        } // Clear div\n        interactionMenu.load(div); // load the new menu\n        div.style.visibility = 'visible';\n      } else {\n        div.style.visibility = 'hidden';\n        this.html.main.focus();\n      }\n    }\n\n    /*\r\n    Section: Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"addEventListener\",\n    value: function addEventListener(eventName, listener) {\n      var _this$eventListeners;\n      (_this$eventListeners = this.eventListeners) === null || _this$eventListeners === void 0 || (_this$eventListeners = _this$eventListeners[eventName]) === null || _this$eventListeners === void 0 || _this$eventListeners.push(listener);\n    }\n  }, {\n    key: \"removeEventListener\",\n    value: function removeEventListener(eventName, listener) {\n      var _this$eventListeners2, _this$eventListeners3;\n      if (!((_this$eventListeners2 = this.eventListeners) !== null && _this$eventListeners2 !== void 0 && _this$eventListeners2[eventName])) return;\n      this.eventListeners[eventName] = (_this$eventListeners3 = this.eventListeners) === null || _this$eventListeners3 === void 0 ? void 0 : _this$eventListeners3[eventName].filter(function (l) {\n        return l !== listener;\n      });\n    }\n  }, {\n    key: \"broadcastEvent\",\n    value: function broadcastEvent(eventName, args) {\n      var _this$eventListeners4;\n      (_this$eventListeners4 = this.eventListeners) === null || _this$eventListeners4 === void 0 || (_this$eventListeners4 = _this$eventListeners4[eventName]) === null || _this$eventListeners4 === void 0 || _this$eventListeners4.forEach(function (eventHandler) {\n        eventHandler(args);\n      });\n    }\n  }, {\n    key: \"onPointerDown\",\n    value: function onPointerDown(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n\n      // Initialise\n      this.playerInput.pointerState = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN;\n      this.playerInput.pointerStart = coords;\n      this.playerInput.pointerCurrent = coords;\n      this.playerInput.pointerEnd = undefined;\n      this.playerInput.pointerTapOrHold = true;\n      this.playerInput.pointerDownDuration = 0;\n      this.html.main.focus();\n      this.broadcastEvent('pointerdown', {\n        coords: coords\n      });\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerMove\",\n    value: function onPointerMove(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerInput.pointerCurrent = coords;\n\n      // If the pointer never moves far from the initial position, then the\n      // pointer interaction is considered a tap or hold.\n      if (this.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN && this.playerInput.pointerTapOrHold) {\n        var distX = this.playerInput.pointerCurrent.x - this.playerInput.pointerStart.x;\n        var distY = this.playerInput.pointerCurrent.y - this.playerInput.pointerStart.y;\n        var pointerDistance = Math.sqrt(distX * distX + distY * distY);\n        if (pointerDistance > _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_DEADZONE_RADIUS) {\n          this.playerInput.pointerTapOrHold = false;\n        }\n      }\n      this.broadcastEvent('pointermove', {\n        coords: coords\n      });\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerUp\",\n    value: function onPointerUp(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      if (this.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN) {\n        this.playerInput.pointerEnd = coords;\n        this.playerInput.pointerState = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.IDLE;\n\n        // Is the pointer action a tap or hold action?\n        if (this.playerInput.pointerTapOrHold) {\n          if (this.playerInput.pointerDownDuration <= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_TAP_DURATION) {\n            this.broadcastEvent('pointertap', {\n              coords: coords\n            });\n          } else {\n            this.broadcastEvent('pointerholdend', {\n              coords: coords,\n              duration: this.playerInput.pointerDownDuration\n            });\n          }\n        }\n      }\n      this.broadcastEvent('pointerup', {\n        coords: coords\n      });\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onKeyDown\",\n    value: function onKeyDown(e) {\n      // Special cases\n      switch (e.key) {\n        // Open home menu\n        case 'Escape':\n          this.setHomeMenu(!this.homeMenu);\n          break;\n\n        // DEBUG\n        case 'x':\n        case 'X':\n          if (!this.interactionMenu) {\n            this.setInteractionMenu(new _avo_interaction__WEBPACK_IMPORTED_MODULE_3__[\"default\"](this));\n          }\n          break;\n        case '-':\n        case '_':\n          this.camera.zoom = Math.max(0.5, this.camera.zoom - 0.5);\n          break;\n        case '+':\n        case '=':\n          this.camera.zoom = Math.min(4, this.camera.zoom + 0.5);\n          break;\n      }\n\n      // General input\n      if (!this.playerInput.keysPressed[e.key]) {\n        this.playerInput.keysPressed[e.key] = {\n          duration: 0,\n          acknowledged: false\n        };\n      }\n      this.broadcastEvent('keydown', {\n        key: e.key\n      });\n    }\n  }, {\n    key: \"onKeyUp\",\n    value: function onKeyUp(e) {\n      var _this$playerInput$key;\n      var duration = ((_this$playerInput$key = this.playerInput.keysPressed[e.key]) === null || _this$playerInput$key === void 0 ? void 0 : _this$playerInput$key.duration) || 0;\n      this.playerInput.keysPressed[e.key] = undefined;\n      this.broadcastEvent('keyup', {\n        key: e.key,\n        duration: duration\n      });\n    }\n  }, {\n    key: \"buttonHome_onClick\",\n    value: function buttonHome_onClick() {\n      this.setHomeMenu(!this.homeMenu);\n    }\n  }, {\n    key: \"buttonFullscreen_onClick\",\n    value: function buttonFullscreen_onClick() {\n      var _this4 = this;\n      var isFullscreen = document.fullscreenElement;\n      if (!isFullscreen) {\n        if (this.html.main.requestFullscreen) {\n          this.html.main.requestFullscreen().then(function () {\n            _this4.html.main.className = 'fullscreen';\n            _this4.updateUI();\n          })[\"catch\"](function (err) {\n            console.error('requestFullscreen() error: ', err);\n          });\n        }\n      } else {\n        var _document$exitFullscr, _document;\n        (_document$exitFullscr = (_document = document).exitFullscreen) === null || _document$exitFullscr === void 0 || _document$exitFullscr.call(_document).then(function () {\n          _this4.html.main.className = '';\n          _this4.updateUI();\n        })[\"catch\"](function (err) {\n          console.error('exitFullscreen() error: ', err);\n        });\n      }\n    }\n  }, {\n    key: \"buttonReload_onClick\",\n    value: function buttonReload_onClick() {\n      var _this$story3;\n      (_this$story3 = this.story) === null || _this$story3 === void 0 || _this$story3.reload();\n    }\n\n    /*\r\n    Section: Gameplay\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"addEntity\",\n    value: function addEntity(entity) {\n      if (!entity) return null;\n      if (!this.entities.includes(entity)) this.entities.push(entity);\n      return entity;\n    }\n  }, {\n    key: \"removeEntity\",\n    value: function removeEntity(entityOrMatchingFn) {\n      if (!entityOrMatchingFn) return;\n      if (typeof entityOrMatchingFn === 'function') {\n        this.entities.filter(entityOrMatchingFn).forEach(function (entity) {\n          entity._expired = true;\n        });\n      } else if (this.entities.includes(entityOrMatchingFn)) {\n        entityOrMatchingFn._expired = true;\n      }\n    }\n  }, {\n    key: \"clearEntities\",\n    value: function clearEntities() {\n      this.entities.forEach(function (entity) {\n        return entity.deconstructor();\n      });\n      this.entities = [];\n    }\n  }, {\n    key: \"resetGameMap\",\n    value: function resetGameMap() {\n      // TODO: do tiles need to be deconstructed?\n      this.gameMap = {\n        // Game map and environment\n        tiles: [],\n        // 2D array of map tiles\n        width: 0,\n        height: 0\n      };\n    }\n  }, {\n    key: \"addRule\",\n    value: function addRule(rule) {\n      if (!rule) return;\n      var id = rule._type;\n      this.rules.set(id, rule);\n    }\n  }, {\n    key: \"clearRules\",\n    value: function clearRules() {\n      var _this5 = this;\n      this.rules.forEach(function (rule, id) {\n        rule.deconstructor();\n        _this5.rules[\"delete\"](id);\n      });\n    }\n  }, {\n    key: \"resetCamera\",\n    value: function resetCamera() {\n      this.camera = {\n        target: null,\n        x: 0,\n        y: 0,\n        zoom: 1\n      };\n    }\n\n    /*\r\n    Section: Painting\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Applies camera transforms to the canvas.\r\n    Should be run right before drawing an Entity (or etc) so the object is drawn\r\n    relative to the camera's view.\r\n     */\n  }, {\n    key: \"applyCameraTransforms\",\n    value: function applyCameraTransforms() {\n      if (this._canvasHasCameraTransforms) throw new Error('Canvas already has camera transforms.');\n      this._canvasHasCameraTransforms = true;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n      c2d.save();\n      c2d.translate(camera.x, camera.y);\n      c2d.scale(camera.zoom, camera.zoom);\n    }\n\n    /*\r\n    Removes camera transforms from the canvas.\r\n     */\n  }, {\n    key: \"undoCameraTransforms\",\n    value: function undoCameraTransforms() {\n      if (!this._canvasHasCameraTransforms) throw new Error('Canvas doesn\\'t have camera transforms.');\n      this._canvasHasCameraTransforms = false;\n      this.canvas2d.restore();\n    }\n\n    /*\r\n    Section: Misc\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"checkCollisions\",\n    value: function checkCollisions() {\n      for (var a = 0; a < this.entities.length; a++) {\n        var entityA = this.entities[a];\n        for (var b = a + 1; b < this.entities.length; b++) {\n          var entityB = this.entities[b];\n          var collisionCorrection = _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(entityA, entityB);\n          if (collisionCorrection) {\n            entityA.onCollision(entityB, collisionCorrection.a);\n            entityB.onCollision(entityA, collisionCorrection.b);\n          }\n        }\n        var range = Math.ceil(entityA.size / _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n        for (var row = entityA.row - range; row <= entityA.row + range; row++) {\n          for (var col = entityA.col - range; col <= entityA.col + range; col++) {\n            var _this$gameMap$tiles;\n            var tile = (_this$gameMap$tiles = this.gameMap.tiles) === null || _this$gameMap$tiles === void 0 || (_this$gameMap$tiles = _this$gameMap$tiles[row]) === null || _this$gameMap$tiles === void 0 ? void 0 : _this$gameMap$tiles[col];\n            var _collisionCorrection = _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(entityA, tile);\n            if (_collisionCorrection) {\n              entityA.onCollision(tile, _collisionCorrection.a);\n              tile.onCollision(entityA, _collisionCorrection.b);\n            }\n          }\n        }\n      }\n    }\n  }, {\n    key: \"resetPlayerInput\",\n    value: function resetPlayerInput() {\n      this.playerInput = {\n        // Pointer (mouse/touchscreen) input\n        // pointerStart/pointerCurrent/pointerEnd = { x, y } \n        pointerState: _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.IDLE,\n        pointerStart: undefined,\n        pointerCurrent: undefined,\n        pointerEnd: undefined,\n        // Pointer metadata\n        pointerTapOrHold: true,\n        // A pointer interaction is a tap or hold if the pointer never travels far from its initial position (i.e. never left the deadzone).\n        pointerDownDuration: 0,\n        // Keyboard input\n        // keysPressed = { key: { duration, acknowledged } }\n        keysPressed: {}\n      };\n    }\n  }]);\n  return AvO;\n}();\n\nfunction getEventCoords(event, element) {\n  var xRatio = element.width && element.offsetWidth ? element.width / element.offsetWidth : 1;\n  var yRatio = element.height && element.offsetHeight ? element.height / element.offsetHeight : 1;\n  var x = event.offsetX * xRatio;\n  var y = event.offsetY * yRatio;\n  return {\n    x: x,\n    y: y\n  };\n}\nfunction stopEvent(e) {\n  if (!e) return false;\n  e.preventDefault && e.preventDefault();\n  e.stopPropagation && e.stopPropagation();\n  e.returnValue = false;\n  e.cancelBubble = true;\n  return false;\n}\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/avo.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ AvO)\n/* harmony export */ });\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\n/* harmony import */ var _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics.js */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_story_examples_example_story_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/story/examples/example-story.js */ \"./src/avo/story/examples/example-story.js\");\n/* harmony import */ var _avo_interaction__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/interaction */ \"./src/avo/interaction/index.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\n\n\n // TODO: remove, or replace with a simpler version\n\nvar searchParams = new URLSearchParams(window.location.search);\nvar DEBUG = searchParams.get('debug') || false;\nvar AvO = /*#__PURE__*/function () {\n  function AvO() {\n    var _this$story;\n    var args = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : {};\n    _classCallCheck(this, AvO);\n    var _args$story = args.story,\n      story = _args$story === void 0 ? _avo_story_examples_example_story_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"] : _args$story,\n      _args$width = args.width,\n      width = _args$width === void 0 ? 24 * _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$width,\n      _args$height = args.height,\n      height = _args$height === void 0 ? 16 * _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE : _args$height;\n    this.html = {\n      main: document.getElementById('main'),\n      canvas: document.getElementById('canvas'),\n      homeMenu: document.getElementById('home-menu'),\n      interactionMenu: document.getElementById('interaction-menu'),\n      buttonHome: document.getElementById('button-home'),\n      buttonFullscreen: document.getElementById('button-fullscreen'),\n      buttonReload: document.getElementById('button-reload')\n    };\n    this.homeMenu = false;\n    this.setHomeMenu(false);\n    this.interactionMenu = false;\n    this.setInteractionMenu(false);\n    this.canvas2d = this.html.canvas.getContext('2d');\n    this.canvasWidth = width;\n    this.canvasHeight = height;\n    this._canvasHasCameraTransforms = false; // Safety check\n\n    this.camera = {\n      target: null,\n      // Target entity to follow. If null, camera is static.\n      x: 0,\n      y: 0,\n      zoom: 1\n    };\n    this.setupUI();\n    this.entities = []; // Game objects\n    this.hero = null; // Main player-controlled entity\n    this.gameMap = {\n      // Game map and environment\n      tiles: [],\n      // 2D array of map tiles\n      width: 0,\n      height: 0\n    };\n    this.rules = new Map(); // Game rules, including win/loss conditions and player controls\n\n    this.story = story ? new story(this) : undefined; // Game story, i.e. the \"container\" that describes an adventure or experience, which contains game objects and game rules and etc\n    this.assets = ((_this$story = this.story) === null || _this$story === void 0 ? void 0 : _this$story.assets) || {};\n    this.secretAssets = {};\n    this.eventListeners = {\n      'keydown': [],\n      'keyup': [],\n      'pointerdown': [],\n      'pointermove': [],\n      'pointerup': [],\n      'pointertap': [],\n      'pointerholdend': []\n    };\n    this.playerInput = {};\n    this.resetPlayerInput();\n    this.initialised = false;\n    this.timeAccumulator = 0;\n    this.prevTime = null;\n    this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n  }\n  _createClass(AvO, [{\n    key: \"initialisationCheck\",\n    value: function initialisationCheck() {\n      var _this = this;\n      // Assets check\n      var allAssetsReady = true;\n      var numReadyAssets = 0;\n      var numTotalAssets = 0;\n      Object.keys(this.assets).forEach(function (id) {\n        var asset = _this.assets[id];\n        allAssetsReady = allAssetsReady && asset.ready;\n        if (asset.ready) numReadyAssets++;\n        numTotalAssets++;\n      });\n      Object.keys(this.secretAssets).forEach(function (id) {\n        var secretAsset = _this.secretAssets[id];\n        var secretAssetIsReady = secretAsset.ready || secretAsset.error;\n        allAssetsReady = allAssetsReady && secretAssetIsReady;\n        if (secretAssetIsReady) numReadyAssets++;\n        numTotalAssets++;\n      });\n\n      // Paint status\n      this.canvas2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      this.canvas2d.textAlign = 'start';\n      this.canvas2d.textBaseline = 'top';\n      this.canvas2d.fillStyle = '#ccc';\n      this.canvas2d.font = \"1em monospace\";\n      this.canvas2d.fillText(\"Loading \".concat(numReadyAssets, \" / \").concat(numTotalAssets, \" \"), _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE, _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n      if (allAssetsReady) {\n        var _this$story2;\n        // Clean up secret assets\n        Object.keys(this.secretAssets).forEach(function (id) {\n          if (_this.secretAssets[id].error) delete _this.secretAssets[id];\n        });\n\n        // Let's go!\n        this.initialised = true;\n        this.showUI();\n        (_this$story2 = this.story) === null || _this$story2 === void 0 || _this$story2.start();\n      }\n    }\n\n    /*\r\n    Section: General Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    The main loop. Run a single frame of gameplay.\r\n    - time: the current/total time (milliseconds) since the game started.\r\n     */\n  }, {\n    key: \"main\",\n    value: function main(time) {\n      var timeStep = this.prevTime ? time - this.prevTime : time;\n      this.prevTime = time;\n      this.timeAccumulator += timeStep;\n\n      // Sanity/safety check: timeStep can be huge\n      // e.g. if player pauses game by switching windows.\n      this.timeAccumulator = Math.min(this.timeAccumulator, _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION * 10);\n      if (this.initialised) {\n        // Keep a consistent \"frame rate\" for logic processing\n        while (this.timeAccumulator >= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION) {\n          this.play();\n          this.timeAccumulator -= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION;\n          // TODO: add safety counter to prevent excessively long while() loops.\n        }\n        // Paint whenever possible\n        this.paint();\n      } else {\n        this.initialisationCheck();\n      }\n      this.nextFrame = window.requestAnimationFrame(this.main.bind(this));\n    }\n\n    /*\r\n    Run the gameplay/physics logic for a single frame.\r\n     */\n  }, {\n    key: \"play\",\n    value: function play() {\n      var _this2 = this;\n      // If a menu is open, pause all action gameplay\n      if (this.homeMenu || this.interactionMenu) return;\n\n      // Run the action gameplay\n      // ----------------\n      this.rules.forEach(function (rule) {\n        return rule.play();\n      });\n      this.entities.forEach(function (entity) {\n        return entity.play();\n      });\n      this.checkCollisions();\n\n      // Cleanup: entities\n      this.entities.filter(function (entity) {\n        return entity._expired;\n      }).forEach(function (entity) {\n        return entity.deconstructor();\n      });\n      this.entities = this.entities.filter(function (entity) {\n        return !entity._expired;\n      });\n\n      // Cleanup: rules\n      this.rules.forEach(function (rule, id) {\n        if (rule.expired) {\n          rule.deconstructor();\n          _this2.rules[\"delete\"](id);\n        }\n      });\n\n      // Sort Entities along the y-axis, for paint()/rendering purposes.\n      // WARNING: inefficient\n      this.entities.sort(function (a, b) {\n        return a.y - b.y;\n      });\n      // ----------------\n\n      // Increment the duration of each currently pressed key\n      Object.keys(this.playerInput.keysPressed).forEach(function (key) {\n        if (_this2.playerInput.keysPressed[key]) _this2.playerInput.keysPressed[key].duration += _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION;\n      });\n\n      // Increment the duration of the pointer being active\n      if (this.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN) {\n        this.playerInput.pointerDownDuration += _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.FRAME_DURATION;\n      }\n    }\n\n    /*\r\n    Paint/draw the game visuals onto the canvas.\r\n     */\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var _this3 = this;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n\n      // Camera Controls: focus the camera on the target entity, if any.\n      // ----------------\n      if (camera.target) {\n        camera.x = this.canvasWidth / 2 - camera.target.x * camera.zoom;\n        camera.y = this.canvasHeight / 2 - camera.target.y * camera.zoom;\n      }\n      c2d.clearRect(0, 0, this.canvasWidth, this.canvasHeight);\n      c2d.resetTransform();\n      c2d.strokeStyle = 'rgba(128, 128, 128, 0.05)';\n      c2d.lineWidth = 2;\n      // ----------------\n\n      // Draw background grid\n      // ----------------\n      var GRID_SIZE = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE * camera.zoom;\n      var offsetX = this.camera.x % GRID_SIZE - GRID_SIZE;\n      var offsetY = this.camera.y % GRID_SIZE - GRID_SIZE;\n      for (var y = offsetY; y < this.canvasHeight; y += GRID_SIZE) {\n        for (var x = offsetX; x < this.canvasWidth; x += GRID_SIZE) {\n          c2d.beginPath();\n          c2d.rect(x, y, GRID_SIZE, GRID_SIZE);\n          c2d.stroke();\n\n          // Debug Grid\n          if (DEBUG) {\n            c2d.fillStyle = '#ccc';\n            c2d.font = \"\".concat(camera.zoom * 0.5, \"em Source Code Pro\");\n            c2d.textAlign = 'center';\n            c2d.textBaseline = 'middle';\n            var col = Math.floor((x - this.camera.x) / GRID_SIZE);\n            var row = Math.floor((y - this.camera.y) / GRID_SIZE);\n            c2d.fillText(col + ',' + row, x + GRID_SIZE / 2, y + GRID_SIZE / 2); // using template strings here messes up colours in Brackets.\n          }\n        }\n      }\n      // ----------------\n\n      // Draw map tiles, entities, and other elements\n      // ----------------\n      var _loop = function _loop(layer) {\n        // 1. draw map tiles first, as these form the background. \n        for (var _row = 0; _row < _this3.gameMap.height; _row++) {\n          for (var _col = 0; _col < _this3.gameMap.width; _col++) {\n            var _this3$gameMap$tiles;\n            (_this3$gameMap$tiles = _this3.gameMap.tiles) === null || _this3$gameMap$tiles === void 0 || (_this3$gameMap$tiles = _this3$gameMap$tiles[_row]) === null || _this3$gameMap$tiles === void 0 || (_this3$gameMap$tiles = _this3$gameMap$tiles[_col]) === null || _this3$gameMap$tiles === void 0 || _this3$gameMap$tiles.paint(layer);\n          }\n        }\n\n        // 2. draw entities next.\n        _this3.entities.forEach(function (entity) {\n          return entity.paint(layer);\n        });\n\n        // 3. draw rules last, as these usually represent HUD visualisations.\n        _this3.rules.forEach(function (rule) {\n          return rule.paint(layer);\n        });\n      };\n      for (var layer = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.MIN_LAYER; layer <= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.MAX_LAYER; layer++) {\n        _loop(layer);\n      }\n      // ----------------\n    }\n\n    /*\r\n    Section: UI\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"setupUI\",\n    value: function setupUI() {\n      this.html.canvas.width = this.canvasWidth;\n      this.html.canvas.height = this.canvasHeight;\n      this.canvas2d.imageSmoothingEnabled = false; /* Pixel art: Maintains sprites' pixel sharpness when scaled up via drawImage() */\n\n      if (window.PointerEvent) {\n        this.html.canvas.addEventListener('pointerdown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('pointermove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('pointerup', this.onPointerUp.bind(this));\n        this.html.canvas.addEventListener('pointercancel', this.onPointerUp.bind(this));\n      } else {\n        this.html.canvas.addEventListener('mousedown', this.onPointerDown.bind(this));\n        this.html.canvas.addEventListener('mousemove', this.onPointerMove.bind(this));\n        this.html.canvas.addEventListener('mouseup', this.onPointerUp.bind(this));\n      }\n\n      // Prevent \"touch and hold to open context menu\" menu on touchscreens.\n      this.html.canvas.addEventListener('touchstart', stopEvent);\n      this.html.canvas.addEventListener('touchmove', stopEvent);\n      this.html.canvas.addEventListener('touchend', stopEvent);\n      this.html.canvas.addEventListener('touchcancel', stopEvent);\n      this.html.buttonHome.addEventListener('click', this.buttonHome_onClick.bind(this));\n      this.html.buttonFullscreen.addEventListener('click', this.buttonFullscreen_onClick.bind(this));\n      this.html.buttonReload.addEventListener('click', this.buttonReload_onClick.bind(this));\n      this.html.main.addEventListener('keydown', this.onKeyDown.bind(this));\n      this.html.main.addEventListener('keyup', this.onKeyUp.bind(this));\n      window.addEventListener('resize', this.updateUI.bind(this));\n      this.updateUI();\n      this.hideUI(); // Hide until all assets are ready\n\n      this.html.main.focus();\n    }\n  }, {\n    key: \"hideUI\",\n    value: function hideUI() {\n      this.html.buttonHome.style.visibility = 'hidden';\n      this.html.buttonReload.style.visibility = 'hidden';\n    }\n  }, {\n    key: \"showUI\",\n    value: function showUI() {\n      this.html.buttonHome.style.visibility = 'visible';\n      this.html.buttonReload.style.visibility = 'visible';\n    }\n  }, {\n    key: \"updateUI\",\n    value: function updateUI() {\n      // Fit the interaction layers (menus, etc) to the canvas\n      var mainDivBounds = this.html.main.getBoundingClientRect();\n      var canvasBounds = this.html.canvas.getBoundingClientRect();\n      this.html.homeMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.homeMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.homeMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.homeMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n      this.html.interactionMenu.style.width = \"\".concat(canvasBounds.width, \"px\");\n      this.html.interactionMenu.style.height = \"\".concat(canvasBounds.height, \"px\");\n      this.html.interactionMenu.style.top = \"\".concat(canvasBounds.top - mainDivBounds.top, \"px\");\n      this.html.interactionMenu.style.left = \"\".concat(canvasBounds.left, \"px\");\n    }\n  }, {\n    key: \"setHomeMenu\",\n    value: function setHomeMenu(homeMenu) {\n      this.homeMenu = homeMenu;\n      if (homeMenu) {\n        this.html.homeMenu.style.visibility = 'visible';\n        this.html.buttonReload.style.visibility = 'hidden';\n      } else {\n        this.html.homeMenu.style.visibility = 'hidden';\n        this.html.buttonReload.style.visibility = 'visible';\n        this.html.main.focus();\n      }\n    }\n  }, {\n    key: \"setInteractionMenu\",\n    value: function setInteractionMenu(interactionMenu) {\n      var div = this.html.interactionMenu;\n      this.interactionMenu && this.interactionMenu.unload(); // Unload the old menu, if any\n      this.interactionMenu = interactionMenu; // Set the new menu\n\n      if (interactionMenu) {\n        while (div.firstChild) {\n          div.removeChild(div.firstChild);\n        } // Clear div\n        interactionMenu.load(div); // load the new menu\n        div.style.visibility = 'visible';\n      } else {\n        div.style.visibility = 'hidden';\n        this.html.main.focus();\n      }\n    }\n\n    /*\r\n    Section: Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"addEventListener\",\n    value: function addEventListener(eventName, listener) {\n      var _this$eventListeners;\n      (_this$eventListeners = this.eventListeners) === null || _this$eventListeners === void 0 || (_this$eventListeners = _this$eventListeners[eventName]) === null || _this$eventListeners === void 0 || _this$eventListeners.push(listener);\n    }\n  }, {\n    key: \"removeEventListener\",\n    value: function removeEventListener(eventName, listener) {\n      var _this$eventListeners2, _this$eventListeners3;\n      if (!((_this$eventListeners2 = this.eventListeners) !== null && _this$eventListeners2 !== void 0 && _this$eventListeners2[eventName])) return;\n      this.eventListeners[eventName] = (_this$eventListeners3 = this.eventListeners) === null || _this$eventListeners3 === void 0 ? void 0 : _this$eventListeners3[eventName].filter(function (l) {\n        return l !== listener;\n      });\n    }\n  }, {\n    key: \"broadcastEvent\",\n    value: function broadcastEvent(eventName, args) {\n      var _this$eventListeners4;\n      (_this$eventListeners4 = this.eventListeners) === null || _this$eventListeners4 === void 0 || (_this$eventListeners4 = _this$eventListeners4[eventName]) === null || _this$eventListeners4 === void 0 || _this$eventListeners4.forEach(function (eventHandler) {\n        eventHandler(args);\n      });\n    }\n  }, {\n    key: \"onPointerDown\",\n    value: function onPointerDown(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n\n      // Initialise\n      this.playerInput.pointerState = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN;\n      this.playerInput.pointerStart = coords;\n      this.playerInput.pointerCurrent = coords;\n      this.playerInput.pointerEnd = undefined;\n      this.playerInput.pointerTapOrHold = true;\n      this.playerInput.pointerDownDuration = 0;\n      this.html.main.focus();\n      this.broadcastEvent('pointerdown', {\n        coords: coords\n      });\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerMove\",\n    value: function onPointerMove(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      this.playerInput.pointerCurrent = coords;\n\n      // If the pointer never moves far from the initial position, then the\n      // pointer interaction is considered a tap or hold.\n      if (this.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN && this.playerInput.pointerTapOrHold) {\n        var distX = this.playerInput.pointerCurrent.x - this.playerInput.pointerStart.x;\n        var distY = this.playerInput.pointerCurrent.y - this.playerInput.pointerStart.y;\n        var pointerDistance = Math.sqrt(distX * distX + distY * distY);\n        if (pointerDistance > _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_DEADZONE_RADIUS) {\n          this.playerInput.pointerTapOrHold = false;\n        }\n      }\n      this.broadcastEvent('pointermove', {\n        coords: coords\n      });\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onPointerUp\",\n    value: function onPointerUp(e) {\n      var coords = getEventCoords(e, this.html.canvas);\n      if (this.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.POINTER_DOWN) {\n        this.playerInput.pointerEnd = coords;\n        this.playerInput.pointerState = _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.IDLE;\n\n        // Is the pointer action a tap or hold action?\n        if (this.playerInput.pointerTapOrHold) {\n          if (this.playerInput.pointerDownDuration <= _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_TAP_DURATION) {\n            this.broadcastEvent('pointertap', {\n              coords: coords\n            });\n          } else {\n            this.broadcastEvent('pointerholdend', {\n              coords: coords,\n              duration: this.playerInput.pointerDownDuration\n            });\n          }\n        }\n      }\n      this.broadcastEvent('pointerup', {\n        coords: coords\n      });\n      return stopEvent(e);\n    }\n  }, {\n    key: \"onKeyDown\",\n    value: function onKeyDown(e) {\n      // Special cases\n      switch (e.key) {\n        // Open home menu\n        case 'Escape':\n          this.setHomeMenu(!this.homeMenu);\n          break;\n\n        // DEBUG\n        case 'x':\n        case 'X':\n          if (!this.interactionMenu) {\n            this.setInteractionMenu(new _avo_interaction__WEBPACK_IMPORTED_MODULE_3__[\"default\"](this));\n          }\n          break;\n        case '-':\n        case '_':\n          this.camera.zoom = Math.max(0.5, this.camera.zoom - 0.5);\n          break;\n        case '+':\n        case '=':\n          this.camera.zoom = Math.min(4, this.camera.zoom + 0.5);\n          break;\n      }\n\n      // General input\n      if (!this.playerInput.keysPressed[e.key]) {\n        this.playerInput.keysPressed[e.key] = {\n          duration: 0,\n          acknowledged: false\n        };\n      }\n      this.broadcastEvent('keydown', {\n        key: e.key\n      });\n    }\n  }, {\n    key: \"onKeyUp\",\n    value: function onKeyUp(e) {\n      var _this$playerInput$key;\n      var duration = ((_this$playerInput$key = this.playerInput.keysPressed[e.key]) === null || _this$playerInput$key === void 0 ? void 0 : _this$playerInput$key.duration) || 0;\n      this.playerInput.keysPressed[e.key] = undefined;\n      this.broadcastEvent('keyup', {\n        key: e.key,\n        duration: duration\n      });\n    }\n  }, {\n    key: \"buttonHome_onClick\",\n    value: function buttonHome_onClick() {\n      this.setHomeMenu(!this.homeMenu);\n    }\n  }, {\n    key: \"buttonFullscreen_onClick\",\n    value: function buttonFullscreen_onClick() {\n      var _this4 = this;\n      var isFullscreen = document.fullscreenElement;\n      if (!isFullscreen) {\n        if (this.html.main.requestFullscreen) {\n          this.html.main.requestFullscreen().then(function () {\n            _this4.html.main.className = 'fullscreen';\n            _this4.updateUI();\n          })[\"catch\"](function (err) {\n            console.error('requestFullscreen() error: ', err);\n          });\n        }\n      } else {\n        var _document$exitFullscr, _document;\n        (_document$exitFullscr = (_document = document).exitFullscreen) === null || _document$exitFullscr === void 0 || _document$exitFullscr.call(_document).then(function () {\n          _this4.html.main.className = '';\n          _this4.updateUI();\n        })[\"catch\"](function (err) {\n          console.error('exitFullscreen() error: ', err);\n        });\n      }\n    }\n  }, {\n    key: \"buttonReload_onClick\",\n    value: function buttonReload_onClick() {\n      var _this$story3;\n      (_this$story3 = this.story) === null || _this$story3 === void 0 || _this$story3.reload();\n    }\n\n    /*\r\n    Section: Gameplay\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"addEntity\",\n    value: function addEntity(entity) {\n      if (!entity) return null;\n      if (!this.entities.includes(entity)) this.entities.push(entity);\n      return entity;\n    }\n  }, {\n    key: \"removeEntity\",\n    value: function removeEntity(entityOrMatchingFn) {\n      if (!entityOrMatchingFn) return;\n      if (typeof entityOrMatchingFn === 'function') {\n        this.entities.filter(entityOrMatchingFn).forEach(function (entity) {\n          entity._expired = true;\n        });\n      } else if (this.entities.includes(entityOrMatchingFn)) {\n        entityOrMatchingFn._expired = true;\n      }\n    }\n  }, {\n    key: \"clearEntities\",\n    value: function clearEntities() {\n      this.entities.forEach(function (entity) {\n        return entity.deconstructor();\n      });\n      this.entities = [];\n    }\n  }, {\n    key: \"resetGameMap\",\n    value: function resetGameMap() {\n      // TODO: do tiles need to be deconstructed?\n      this.gameMap = {\n        // Game map and environment\n        tiles: [],\n        // 2D array of map tiles\n        width: 0,\n        height: 0\n      };\n    }\n  }, {\n    key: \"addRule\",\n    value: function addRule(rule) {\n      if (!rule) return;\n      var id = rule._type;\n      this.rules.set(id, rule);\n    }\n  }, {\n    key: \"clearRules\",\n    value: function clearRules() {\n      var _this5 = this;\n      this.rules.forEach(function (rule, id) {\n        rule.deconstructor();\n        _this5.rules[\"delete\"](id);\n      });\n    }\n  }, {\n    key: \"resetCamera\",\n    value: function resetCamera() {\n      this.camera = {\n        target: null,\n        x: 0,\n        y: 0,\n        zoom: 1\n      };\n    }\n\n    /*\r\n    Section: Painting\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Applies camera transforms to the canvas.\r\n    Should be run right before drawing an Entity (or etc) so the object is drawn\r\n    relative to the camera's view.\r\n     */\n  }, {\n    key: \"applyCameraTransforms\",\n    value: function applyCameraTransforms() {\n      if (this._canvasHasCameraTransforms) throw new Error('Canvas already has camera transforms.');\n      this._canvasHasCameraTransforms = true;\n      var c2d = this.canvas2d;\n      var camera = this.camera;\n      c2d.save();\n      c2d.translate(camera.x, camera.y);\n      c2d.scale(camera.zoom, camera.zoom);\n    }\n\n    /*\r\n    Removes camera transforms from the canvas.\r\n     */\n  }, {\n    key: \"undoCameraTransforms\",\n    value: function undoCameraTransforms() {\n      if (!this._canvasHasCameraTransforms) throw new Error('Canvas doesn\\'t have camera transforms.');\n      this._canvasHasCameraTransforms = false;\n      this.canvas2d.restore();\n    }\n\n    /*\r\n    Section: Misc\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"checkCollisions\",\n    value: function checkCollisions() {\n      for (var a = 0; a < this.entities.length; a++) {\n        var entityA = this.entities[a];\n        for (var b = a + 1; b < this.entities.length; b++) {\n          var entityB = this.entities[b];\n          var collisionCorrection = _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(entityA, entityB);\n          if (collisionCorrection) {\n            entityA.onCollision(entityB, collisionCorrection.a);\n            entityB.onCollision(entityA, collisionCorrection.b);\n          }\n        }\n        var range = Math.ceil(entityA.size / _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.TILE_SIZE);\n        for (var row = entityA.row - range; row <= entityA.row + range; row++) {\n          for (var col = entityA.col - range; col <= entityA.col + range; col++) {\n            var _this$gameMap$tiles;\n            var tile = (_this$gameMap$tiles = this.gameMap.tiles) === null || _this$gameMap$tiles === void 0 || (_this$gameMap$tiles = _this$gameMap$tiles[row]) === null || _this$gameMap$tiles === void 0 ? void 0 : _this$gameMap$tiles[col];\n            var _collisionCorrection = _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].checkCollision(entityA, tile);\n            if (_collisionCorrection) {\n              entityA.onCollision(tile, _collisionCorrection.a);\n              tile.onCollision(entityA, _collisionCorrection.b);\n            }\n          }\n        }\n      }\n    }\n  }, {\n    key: \"resetPlayerInput\",\n    value: function resetPlayerInput() {\n      this.playerInput = {\n        // Pointer (mouse/touchscreen) input\n        // pointerStart/pointerCurrent/pointerEnd = { x, y } \n        pointerState: _avo_constants_js__WEBPACK_IMPORTED_MODULE_0__.POINTER_STATES.IDLE,\n        pointerStart: undefined,\n        pointerCurrent: undefined,\n        pointerEnd: undefined,\n        // Pointer metadata\n        pointerTapOrHold: true,\n        // A pointer interaction is a tap or hold if the pointer never travels far from its initial position (i.e. never left the deadzone).\n        pointerDownDuration: 0,\n        // Keyboard input\n        // keysPressed = { key: { duration, acknowledged } }\n        keysPressed: {}\n      };\n    }\n  }]);\n  return AvO;\n}();\n\nfunction getEventCoords(event, element) {\n  var xRatio = element.width && element.offsetWidth ? element.width / element.offsetWidth : 1;\n  var yRatio = element.height && element.offsetHeight ? element.height / element.offsetHeight : 1;\n  var x = event.offsetX * xRatio;\n  var y = event.offsetY * yRatio;\n  return {\n    x: x,\n    y: y\n  };\n}\nfunction stopEvent(e) {\n  if (!e) return false;\n  e.preventDefault && e.preventDefault();\n  e.stopPropagation && e.stopPropagation();\n  e.returnValue = false;\n  e.cancelBubble = true;\n  return false;\n}\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/avo.js?");
 
 /***/ }),
 
@@ -40,6 +40,56 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/avo/entity/examples/ball.js":
+/*!*****************************************!*\
+  !*** ./src/avo/entity/examples/ball.js ***!
+  \*****************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Ball)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar Ball = /*#__PURE__*/function (_Entity) {\n  _inherits(Ball, _Entity);\n  var _super = _createSuper(Ball);\n  function Ball(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    _classCallCheck(this, Ball);\n    _this = _super.call(this, app);\n    _this._type = 'ball';\n    _this.colour = '#48c';\n    _this.solid = true;\n    _this.movable = true;\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    return _this;\n  }\n  return _createClass(Ball);\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/examples/ball.js?");
+
+/***/ }),
+
+/***/ "./src/avo/entity/examples/enemy.js":
+/*!******************************************!*\
+  !*** ./src/avo/entity/examples/enemy.js ***!
+  \******************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Enemy)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_entity_examples_shot_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/entity/examples/shot.js */ \"./src/avo/entity/examples/shot.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\nvar Enemy = /*#__PURE__*/function (_Entity) {\n  _inherits(Enemy, _Entity);\n  var _super = _createSuper(Enemy);\n  function Enemy(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    _classCallCheck(this, Enemy);\n    _this = _super.call(this, app);\n    _this._type = 'enemy';\n    _this.colour = '#4c4';\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE / 2;\n    _this.action = {\n      name: 'shoot',\n      counter: 0,\n      acknowledged: false\n    };\n    return _this;\n  }\n  _createClass(Enemy, [{\n    key: \"play\",\n    value: function play() {\n      _get(_getPrototypeOf(Enemy.prototype), \"play\", this).call(this);\n      this.processAction();\n    }\n  }, {\n    key: \"processAction\",\n    value: function processAction() {\n      if (!this.action) return;\n      var action = this.action;\n      var app = this._app;\n      if (action.name === 'shoot') {\n        var DURATION = 500;\n        var progress = action.counter / DURATION;\n        if (!action.acknowledged && progress > 0.5) {\n          action.acknowledged = true;\n          app.addEntity(new _avo_entity_examples_shot_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"](app, this.x, this.y, this.rotation, this));\n        }\n        action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.FRAME_DURATION;\n        if (action.counter >= DURATION) {\n          this.action = {\n            name: 'shoot',\n            counter: 0,\n            acknowledged: false\n          };\n        }\n      }\n    }\n  }]);\n  return Enemy;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/examples/enemy.js?");
+
+/***/ }),
+
+/***/ "./src/avo/entity/examples/hero.js":
+/*!*****************************************!*\
+  !*** ./src/avo/entity/examples/hero.js ***!
+  \*****************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Hero)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction ownKeys(e, r) { var t = Object.keys(e); if (Object.getOwnPropertySymbols) { var o = Object.getOwnPropertySymbols(e); r && (o = o.filter(function (r) { return Object.getOwnPropertyDescriptor(e, r).enumerable; })), t.push.apply(t, o); } return t; }\nfunction _objectSpread(e) { for (var r = 1; r < arguments.length; r++) { var t = null != arguments[r] ? arguments[r] : {}; r % 2 ? ownKeys(Object(t), !0).forEach(function (r) { _defineProperty(e, r, t[r]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(t)) : ownKeys(Object(t)).forEach(function (r) { Object.defineProperty(e, r, Object.getOwnPropertyDescriptor(t, r)); }); } return e; }\nfunction _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar INVULNERABILITY_WINDOW = 3000;\nvar MOVE_ACTION_CYCLE_DURATION = 500;\nvar Hero = /*#__PURE__*/function (_Entity) {\n  _inherits(Hero, _Entity);\n  var _super = _createSuper(Hero);\n  function Hero(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    _classCallCheck(this, Hero);\n    _this = _super.call(this, app);\n    _this._type = 'hero';\n    _this.colour = '#000';\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.intent = undefined;\n    _this.action = undefined;\n    _this.health = 3;\n    _this.invulnerability = 0; // Invulnerability time\n    return _this;\n  }\n\n  /*\r\n  Section: General Logic\r\n  ----------------------------------------------------------------------------\r\n   */\n  _createClass(Hero, [{\n    key: \"play\",\n    value: function play() {\n      _get(_getPrototypeOf(Hero.prototype), \"play\", this).call(this);\n      this.processIntent();\n      this.processAction();\n\n      // Count down invulnerability time\n      if (this.invulnerability > 0) {\n        this.invulnerability = Math.max(this.invulnerability - _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION, 0);\n      }\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n      if (this.invulnerability > 0) {\n        // If invulnerable, flash!\n        var flash = Math.floor(this.invulnerability / 300) % 2;\n        if (flash === 1) return;\n      }\n      this.colour = app.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.POINTER_STATES.POINTER_DOWN ? '#e42' : '#c44';\n      _get(_getPrototypeOf(Hero.prototype), \"paint\", this).call(this, layer);\n      var c2d = app.canvas2d;\n      var animationSpriteSheet = app.assets['hero-4dir'];\n      if (!animationSpriteSheet) return;\n      this._app.applyCameraTransforms();\n      var SPRITE_SIZE = 48;\n      var FLIP_SPRITE = 1;\n      var SPRITE_SCALE = 2 * (Math.min(this.health / 3, 1) * 0.5 + 0.5); // Shrink after taking damage\n\n      // Draw any special vfx\n      var action = this.action;\n      if ((action === null || action === void 0 ? void 0 : action.name) === 'dash' && (action === null || action === void 0 ? void 0 : action.state) === 'execution' && layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.LAYERS.MIDDLE) {\n        // Draw a \"dash line\"\n        var dashLength = this.size * 2;\n        var dashWidth = this.size;\n        var dashTailX = this.x - dashLength * Math.cos(this.rotation);\n        var dashTailY = this.y - dashLength * Math.sin(this.rotation);\n        c2d.beginPath();\n        c2d.moveTo(this.x, this.y);\n        c2d.lineTo(dashTailX, dashTailY);\n        c2d.strokeStyle = 'rgba(255, 255, 0, 0.5)';\n        c2d.lineWidth = dashWidth;\n        c2d.stroke();\n      }\n\n      // Draw the sprite\n      if (layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.LAYERS.MIDDLE) {\n        var srcX = this.getSpriteCol() * SPRITE_SIZE;\n        var srcY = this.getSpriteRow() * SPRITE_SIZE;\n        var sizeX = SPRITE_SIZE;\n        var sizeY = SPRITE_SIZE;\n        c2d.translate(this.x, this.y); // 1. This moves the 'drawing origin' to match the position of (the centre of) the Entity.\n        c2d.scale(SPRITE_SCALE * FLIP_SPRITE, SPRITE_SCALE); // 2. This ensures the sprite scales with the 'drawing origin' as the anchor point.\n        // c2d.rotate(this.rotation)  // 3. If we wanted to, we could rotate the sprite around the 'drawing origin'.\n\n        // 4. tgtX and tgtY specify where to draw the sprite, relative to the 'drawing origin'.\n        var tgtX = -sizeX / 2; // Align centre of sprite to origin\n        var tgtY = -sizeY * 0.75; // Align bottom(-ish) of sprite to origin\n\n        c2d.drawImage(animationSpriteSheet.img, srcX, srcY, sizeX, sizeY, tgtX, tgtY, sizeX, sizeY);\n      }\n      this._app.undoCameraTransforms();\n    }\n\n    /*\r\n    Section: Game Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"applyEffect\",\n    value: function applyEffect(effect, source) {\n      _get(_getPrototypeOf(Hero.prototype), \"applyEffect\", this).call(this, effect, source);\n      if (!effect) return;\n      if (effect.name === 'damage') {\n        if (this.invulnerability === 0) {\n          this.health = Math.max(this.health - 1, 0);\n          this.invulnerability = INVULNERABILITY_WINDOW;\n        }\n      }\n    }\n\n    /*\r\n    Section: Intent and Actions\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Translate intent into action.\r\n     */\n  }, {\n    key: \"processIntent\",\n    value: function processIntent() {\n      // Failsafe\n      if (!this.action) this.goIdle();\n      var action = this.action;\n      var intent = this.intent;\n      if (!intent) {\n        // Go idle\n        if ((action === null || action === void 0 ? void 0 : action.name) === 'move') this.goIdle();\n      } else {\n        // Perform a new action\n        // Note: every 'move' action is considered a new action\n\n        if ((action === null || action === void 0 ? void 0 : action.name) === 'idle' || (action === null || action === void 0 ? void 0 : action.name) === 'move') {\n          // Can the action be overwritten by a new action? If not, the action must play through to its finish.\n          this.action = _objectSpread(_objectSpread({}, intent), {}, {\n            name: intent.name,\n            counter: action.name === intent.name ? action.counter : 0 // If the current action and new intent have the same name, it's just a continuation of the idle or move action, but with other new values (e.g. new directions)\n          });\n        }\n      }\n    }\n\n    /*\r\n    Perform the action.\r\n     */\n  }, {\n    key: \"processAction\",\n    value: function processAction() {\n      if (!this.action) return;\n      var action = this.action;\n      if (action.name === 'idle') {\n\n        // Do nothing\n      } else if (action.name === 'move') {\n        var moveAcceleration = this.moveAcceleration || 0;\n        var directionX = action.directionX || 0;\n        var directionY = action.directionY || 0;\n        var actionRotation = Math.atan2(directionY, directionX);\n        this.moveX += moveAcceleration * Math.cos(actionRotation);\n        this.moveY += moveAcceleration * Math.sin(actionRotation);\n        this.rotation = actionRotation;\n        action.counter = (action.counter + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION) % MOVE_ACTION_CYCLE_DURATION;\n      } else if (action.name === 'dash') {\n        var WINDUP_DURATION = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION * 5;\n        var EXECUTION_DURATION = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION * 2;\n        var WINDDOWN_DURATION = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION * 10;\n        var PUSH_POWER = this.size * 0.3;\n        var MAX_PUSH = EXECUTION_DURATION / 1000 * 60 * PUSH_POWER;\n        if (!action.state) {\n          // Trigger only once, at the start of the action\n\n          // Figure out the initial direction of the dash\n          var _directionX = action.directionX || 0;\n          var _directionY = action.directionY || 0;\n          this.rotation = _directionX === 0 && _directionY === 0 ? this.rotation : Math.atan2(_directionY, _directionX);\n          action.rotation = this.rotation; // Records the initial direction of the dash\n\n          action.state = 'windup';\n        }\n        if (action.state === 'windup') {\n          action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION;\n          if (action.counter >= WINDUP_DURATION) {\n            action.state = 'execution';\n            action.counter = 0;\n          }\n        } else if (action.state === 'execution') {\n          this.pushX += PUSH_POWER * Math.cos(action.rotation);\n          this.pushY += PUSH_POWER * Math.sin(action.rotation);\n          action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION;\n          if (action.counter >= EXECUTION_DURATION) {\n            action.state = 'winddown';\n            action.counter = 0;\n          }\n        } else if (action.state === 'winddown') {\n          action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION;\n          if (action.counter >= WINDDOWN_DURATION) {\n            this.goIdle();\n          }\n        }\n      }\n    }\n  }, {\n    key: \"goIdle\",\n    value: function goIdle() {\n      this.action = {\n        name: 'idle',\n        counter: 0\n      };\n    }\n\n    /*\r\n    Section: Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Triggers when this entity hits/touches/intersects with another.\r\n     */\n  }, {\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(Hero.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n      if (!target) return;\n    }\n\n    /*\r\n    Section: Physics/Getters and Setters\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"moveDeceleration\",\n    get: function get() {\n      var _this$action;\n      if (((_this$action = this.action) === null || _this$action === void 0 ? void 0 : _this$action.name) === 'move') return 0;\n      return this._moveDeceleration;\n    }\n  }, {\n    key: \"pushDeceleration\",\n    get: function get() {\n      var _this$action2, _this$action3;\n      if (((_this$action2 = this.action) === null || _this$action2 === void 0 ? void 0 : _this$action2.name) === 'dash' && ((_this$action3 = this.action) === null || _this$action3 === void 0 ? void 0 : _this$action3.state) === 'execution') return 0;\n      return this._pushDeceleration;\n    }\n\n    /*\r\n    Section: Animation\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"getSpriteCol\",\n    value: function getSpriteCol() {\n      switch (this.getSpriteDirection()) {\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.NORTH:\n          return 1;\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.EAST:\n          return 2;\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.SOUTH:\n          return 0;\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.WEST:\n          return 3;\n      }\n      return 0;\n    }\n  }, {\n    key: \"getSpriteRow\",\n    value: function getSpriteRow() {\n      var action = this.action;\n      if (!action) return 0;\n      if (action.name === 'move') {\n        var progress = action.counter / MOVE_ACTION_CYCLE_DURATION;\n        if (progress < 0.3) return 2;else if (progress < 0.5) return 1;else if (progress < 0.8) return 3;else if (progress < 1) return 1;\n      } else if (action.name === 'dash') {\n        if (action.state === 'windup') return 4;else if (action.state === 'execution') return 1;else if (action.state === 'winddown') return 1;\n      }\n      return 0;\n    }\n  }]);\n  return Hero;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/examples/hero.js?");
+
+/***/ }),
+
+/***/ "./src/avo/entity/examples/shot.js":
+/*!*****************************************!*\
+  !*** ./src/avo/entity/examples/shot.js ***!
+  \*****************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Shot)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar Shot = /*#__PURE__*/function (_Entity) {\n  _inherits(Shot, _Entity);\n  var _super = _createSuper(Shot);\n  function Shot(app) {\n    var _this;\n    var x = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var y = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var rotation = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 0;\n    var source = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : undefined;\n    _classCallCheck(this, Shot);\n    _this = _super.call(this, app);\n    _this._type = 'shot';\n    _this.source = source; // The entity that this shot came from\n\n    _this.colour = '#cc4';\n    _this.solid = false;\n    _this.movable = true;\n    _this.size = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.x = x;\n    _this.y = y;\n    _this.rotation = rotation;\n    _this.moveAcceleration = 1;\n    _this.moveDeceleration = 0;\n    _this.moveMaxSpeed = _this.size * 2;\n    _this.distance = _this.size * 1;\n    return _this;\n  }\n  _createClass(Shot, [{\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(Shot.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n      if (this.source !== target && target.solid) {\n        target.applyEffect({\n          name: 'damage'\n        }, this);\n        this._expired = true;\n      }\n    }\n  }, {\n    key: \"play\",\n    value: function play() {\n      _get(_getPrototypeOf(Shot.prototype), \"play\", this).call(this);\n      var moveAcceleration = this.moveAcceleration || 0;\n      this.moveX += moveAcceleration * Math.cos(this.rotation);\n      this.moveY += moveAcceleration * Math.sin(this.rotation);\n      this.distance -= moveAcceleration;\n      if (this.distance <= 0) this._expired = true;\n    }\n  }]);\n  return Shot;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/examples/shot.js?");
+
+/***/ }),
+
+/***/ "./src/avo/entity/examples/wall.js":
+/*!*****************************************!*\
+  !*** ./src/avo/entity/examples/wall.js ***!
+  \*****************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Wall)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar Wall = /*#__PURE__*/function (_Entity) {\n  _inherits(Wall, _Entity);\n  var _super = _createSuper(Wall);\n  function Wall(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var width = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 1;\n    var height = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 1;\n    var cutCorner = arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : false;\n    _classCallCheck(this, Wall);\n    _this = _super.call(this, app);\n    _this._type = 'wall';\n    _this.colour = '#888';\n    _this.solid = true;\n    _this.movable = false;\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.shape = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.SHAPES.POLYGON;\n    _this.shapePolygonPath = [];\n    if (cutCorner !== 'nw') _this.shapePolygonPath.push(0, 0);\n    if (cutCorner !== 'ne') _this.shapePolygonPath.push(width * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, 0);\n    if (cutCorner !== 'se') _this.shapePolygonPath.push(width * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, height * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    if (cutCorner !== 'sw') _this.shapePolygonPath.push(0, height * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    return _this;\n  }\n  return _createClass(Wall);\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/examples/wall.js?");
+
+/***/ }),
+
 /***/ "./src/avo/entity/index.js":
 /*!*********************************!*\
   !*** ./src/avo/entity/index.js ***!
@@ -47,56 +97,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var _entity_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./entity.js */ \"./src/avo/entity/entity.js\");\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (_entity_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/index.js?");
-
-/***/ }),
-
-/***/ "./src/avo/entity/types/ball.js":
-/*!**************************************!*\
-  !*** ./src/avo/entity/types/ball.js ***!
-  \**************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Ball)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar Ball = /*#__PURE__*/function (_Entity) {\n  _inherits(Ball, _Entity);\n  var _super = _createSuper(Ball);\n  function Ball(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    _classCallCheck(this, Ball);\n    _this = _super.call(this, app);\n    _this._type = 'ball';\n    _this.colour = '#48c';\n    _this.solid = true;\n    _this.movable = true;\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    return _this;\n  }\n  return _createClass(Ball);\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/types/ball.js?");
-
-/***/ }),
-
-/***/ "./src/avo/entity/types/enemy.js":
-/*!***************************************!*\
-  !*** ./src/avo/entity/types/enemy.js ***!
-  \***************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Enemy)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_entity_types_shot_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/entity/types/shot.js */ \"./src/avo/entity/types/shot.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\nvar Enemy = /*#__PURE__*/function (_Entity) {\n  _inherits(Enemy, _Entity);\n  var _super = _createSuper(Enemy);\n  function Enemy(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    _classCallCheck(this, Enemy);\n    _this = _super.call(this, app);\n    _this._type = 'enemy';\n    _this.colour = '#4c4';\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE / 2;\n    _this.action = {\n      name: 'shoot',\n      counter: 0,\n      acknowledged: false\n    };\n    return _this;\n  }\n  _createClass(Enemy, [{\n    key: \"play\",\n    value: function play() {\n      _get(_getPrototypeOf(Enemy.prototype), \"play\", this).call(this);\n      this.processAction();\n    }\n  }, {\n    key: \"processAction\",\n    value: function processAction() {\n      if (!this.action) return;\n      var action = this.action;\n      var app = this._app;\n      if (action.name === 'shoot') {\n        var DURATION = 500;\n        var progress = action.counter / DURATION;\n        if (!action.acknowledged && progress > 0.5) {\n          action.acknowledged = true;\n          app.addEntity(new _avo_entity_types_shot_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"](app, this.x, this.y, this.rotation, this));\n        }\n        action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.FRAME_DURATION;\n        if (action.counter >= DURATION) {\n          this.action = {\n            name: 'shoot',\n            counter: 0,\n            acknowledged: false\n          };\n        }\n      }\n    }\n  }]);\n  return Enemy;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/types/enemy.js?");
-
-/***/ }),
-
-/***/ "./src/avo/entity/types/hero.js":
-/*!**************************************!*\
-  !*** ./src/avo/entity/types/hero.js ***!
-  \**************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Hero)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction ownKeys(e, r) { var t = Object.keys(e); if (Object.getOwnPropertySymbols) { var o = Object.getOwnPropertySymbols(e); r && (o = o.filter(function (r) { return Object.getOwnPropertyDescriptor(e, r).enumerable; })), t.push.apply(t, o); } return t; }\nfunction _objectSpread(e) { for (var r = 1; r < arguments.length; r++) { var t = null != arguments[r] ? arguments[r] : {}; r % 2 ? ownKeys(Object(t), !0).forEach(function (r) { _defineProperty(e, r, t[r]); }) : Object.getOwnPropertyDescriptors ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(t)) : ownKeys(Object(t)).forEach(function (r) { Object.defineProperty(e, r, Object.getOwnPropertyDescriptor(t, r)); }); } return e; }\nfunction _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar INVULNERABILITY_WINDOW = 3000;\nvar MOVE_ACTION_CYCLE_DURATION = 500;\nvar Hero = /*#__PURE__*/function (_Entity) {\n  _inherits(Hero, _Entity);\n  var _super = _createSuper(Hero);\n  function Hero(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    _classCallCheck(this, Hero);\n    _this = _super.call(this, app);\n    _this._type = 'hero';\n    _this.colour = '#000';\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.intent = undefined;\n    _this.action = undefined;\n    _this.health = 3;\n    _this.invulnerability = 0; // Invulnerability time\n    return _this;\n  }\n\n  /*\r\n  Section: General Logic\r\n  ----------------------------------------------------------------------------\r\n   */\n  _createClass(Hero, [{\n    key: \"play\",\n    value: function play() {\n      _get(_getPrototypeOf(Hero.prototype), \"play\", this).call(this);\n      this.processIntent();\n      this.processAction();\n\n      // Count down invulnerability time\n      if (this.invulnerability > 0) {\n        this.invulnerability = Math.max(this.invulnerability - _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION, 0);\n      }\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var app = this._app;\n      if (this.invulnerability > 0) {\n        // If invulnerable, flash!\n        var flash = Math.floor(this.invulnerability / 300) % 2;\n        if (flash === 1) return;\n      }\n      this.colour = app.playerInput.pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.POINTER_STATES.POINTER_DOWN ? '#e42' : '#c44';\n      _get(_getPrototypeOf(Hero.prototype), \"paint\", this).call(this, layer);\n      var c2d = app.canvas2d;\n      var animationSpriteSheet = app.assets['hero-4dir'];\n      if (!animationSpriteSheet) return;\n      this._app.applyCameraTransforms();\n      var SPRITE_SIZE = 48;\n      var FLIP_SPRITE = 1;\n      var SPRITE_SCALE = 2 * (Math.min(this.health / 3, 1) * 0.5 + 0.5); // Shrink after taking damage\n\n      // Draw any special vfx\n      var action = this.action;\n      if ((action === null || action === void 0 ? void 0 : action.name) === 'dash' && (action === null || action === void 0 ? void 0 : action.state) === 'execution' && layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.LAYERS.MIDDLE) {\n        // Draw a \"dash line\"\n        var dashLength = this.size * 2;\n        var dashWidth = this.size;\n        var dashTailX = this.x - dashLength * Math.cos(this.rotation);\n        var dashTailY = this.y - dashLength * Math.sin(this.rotation);\n        c2d.beginPath();\n        c2d.moveTo(this.x, this.y);\n        c2d.lineTo(dashTailX, dashTailY);\n        c2d.strokeStyle = 'rgba(255, 255, 0, 0.5)';\n        c2d.lineWidth = dashWidth;\n        c2d.stroke();\n      }\n\n      // Draw the sprite\n      if (layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.LAYERS.MIDDLE) {\n        var srcX = this.getSpriteCol() * SPRITE_SIZE;\n        var srcY = this.getSpriteRow() * SPRITE_SIZE;\n        var sizeX = SPRITE_SIZE;\n        var sizeY = SPRITE_SIZE;\n        c2d.translate(this.x, this.y); // 1. This moves the 'drawing origin' to match the position of (the centre of) the Entity.\n        c2d.scale(SPRITE_SCALE * FLIP_SPRITE, SPRITE_SCALE); // 2. This ensures the sprite scales with the 'drawing origin' as the anchor point.\n        // c2d.rotate(this.rotation)  // 3. If we wanted to, we could rotate the sprite around the 'drawing origin'.\n\n        // 4. tgtX and tgtY specify where to draw the sprite, relative to the 'drawing origin'.\n        var tgtX = -sizeX / 2; // Align centre of sprite to origin\n        var tgtY = -sizeY * 0.75; // Align bottom(-ish) of sprite to origin\n\n        c2d.drawImage(animationSpriteSheet.img, srcX, srcY, sizeX, sizeY, tgtX, tgtY, sizeX, sizeY);\n      }\n      this._app.undoCameraTransforms();\n    }\n\n    /*\r\n    Section: Game Logic\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"applyEffect\",\n    value: function applyEffect(effect, source) {\n      _get(_getPrototypeOf(Hero.prototype), \"applyEffect\", this).call(this, effect, source);\n      if (!effect) return;\n      if (effect.name === 'damage') {\n        if (this.invulnerability === 0) {\n          this.health = Math.max(this.health - 1, 0);\n          this.invulnerability = INVULNERABILITY_WINDOW;\n        }\n      }\n    }\n\n    /*\r\n    Section: Intent and Actions\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Translate intent into action.\r\n     */\n  }, {\n    key: \"processIntent\",\n    value: function processIntent() {\n      // Failsafe\n      if (!this.action) this.goIdle();\n      var action = this.action;\n      var intent = this.intent;\n      if (!intent) {\n        // Go idle\n        if ((action === null || action === void 0 ? void 0 : action.name) === 'move') this.goIdle();\n      } else {\n        // Perform a new action\n        // Note: every 'move' action is considered a new action\n\n        if ((action === null || action === void 0 ? void 0 : action.name) === 'idle' || (action === null || action === void 0 ? void 0 : action.name) === 'move') {\n          // Can the action be overwritten by a new action? If not, the action must play through to its finish.\n          this.action = _objectSpread(_objectSpread({}, intent), {}, {\n            name: intent.name,\n            counter: action.name === intent.name ? action.counter : 0 // If the current action and new intent have the same name, it's just a continuation of the idle or move action, but with other new values (e.g. new directions)\n          });\n        }\n      }\n    }\n\n    /*\r\n    Perform the action.\r\n     */\n  }, {\n    key: \"processAction\",\n    value: function processAction() {\n      if (!this.action) return;\n      var action = this.action;\n      if (action.name === 'idle') {\n\n        // Do nothing\n      } else if (action.name === 'move') {\n        var moveAcceleration = this.moveAcceleration || 0;\n        var directionX = action.directionX || 0;\n        var directionY = action.directionY || 0;\n        var actionRotation = Math.atan2(directionY, directionX);\n        this.moveX += moveAcceleration * Math.cos(actionRotation);\n        this.moveY += moveAcceleration * Math.sin(actionRotation);\n        this.rotation = actionRotation;\n        action.counter = (action.counter + _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION) % MOVE_ACTION_CYCLE_DURATION;\n      } else if (action.name === 'dash') {\n        var WINDUP_DURATION = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION * 5;\n        var EXECUTION_DURATION = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION * 2;\n        var WINDDOWN_DURATION = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION * 10;\n        var PUSH_POWER = this.size * 0.3;\n        var MAX_PUSH = EXECUTION_DURATION / 1000 * 60 * PUSH_POWER;\n        if (!action.state) {\n          // Trigger only once, at the start of the action\n\n          // Figure out the initial direction of the dash\n          var _directionX = action.directionX || 0;\n          var _directionY = action.directionY || 0;\n          this.rotation = _directionX === 0 && _directionY === 0 ? this.rotation : Math.atan2(_directionY, _directionX);\n          action.rotation = this.rotation; // Records the initial direction of the dash\n\n          action.state = 'windup';\n        }\n        if (action.state === 'windup') {\n          action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION;\n          if (action.counter >= WINDUP_DURATION) {\n            action.state = 'execution';\n            action.counter = 0;\n          }\n        } else if (action.state === 'execution') {\n          this.pushX += PUSH_POWER * Math.cos(action.rotation);\n          this.pushY += PUSH_POWER * Math.sin(action.rotation);\n          action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION;\n          if (action.counter >= EXECUTION_DURATION) {\n            action.state = 'winddown';\n            action.counter = 0;\n          }\n        } else if (action.state === 'winddown') {\n          action.counter += _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.FRAME_DURATION;\n          if (action.counter >= WINDDOWN_DURATION) {\n            this.goIdle();\n          }\n        }\n      }\n    }\n  }, {\n    key: \"goIdle\",\n    value: function goIdle() {\n      this.action = {\n        name: 'idle',\n        counter: 0\n      };\n    }\n\n    /*\r\n    Section: Event Handling\r\n    ----------------------------------------------------------------------------\r\n     */\n\n    /*\r\n    Triggers when this entity hits/touches/intersects with another.\r\n     */\n  }, {\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(Hero.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n      if (!target) return;\n    }\n\n    /*\r\n    Section: Physics/Getters and Setters\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"moveDeceleration\",\n    get: function get() {\n      var _this$action;\n      if (((_this$action = this.action) === null || _this$action === void 0 ? void 0 : _this$action.name) === 'move') return 0;\n      return this._moveDeceleration;\n    }\n  }, {\n    key: \"pushDeceleration\",\n    get: function get() {\n      var _this$action2, _this$action3;\n      if (((_this$action2 = this.action) === null || _this$action2 === void 0 ? void 0 : _this$action2.name) === 'dash' && ((_this$action3 = this.action) === null || _this$action3 === void 0 ? void 0 : _this$action3.state) === 'execution') return 0;\n      return this._pushDeceleration;\n    }\n\n    /*\r\n    Section: Animation\r\n    ----------------------------------------------------------------------------\r\n     */\n  }, {\n    key: \"getSpriteCol\",\n    value: function getSpriteCol() {\n      switch (this.getSpriteDirection()) {\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.NORTH:\n          return 1;\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.EAST:\n          return 2;\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.SOUTH:\n          return 0;\n        case _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.DIRECTIONS.WEST:\n          return 3;\n      }\n      return 0;\n    }\n  }, {\n    key: \"getSpriteRow\",\n    value: function getSpriteRow() {\n      var action = this.action;\n      if (!action) return 0;\n      if (action.name === 'move') {\n        var progress = action.counter / MOVE_ACTION_CYCLE_DURATION;\n        if (progress < 0.3) return 2;else if (progress < 0.5) return 1;else if (progress < 0.8) return 3;else if (progress < 1) return 1;\n      } else if (action.name === 'dash') {\n        if (action.state === 'windup') return 4;else if (action.state === 'execution') return 1;else if (action.state === 'winddown') return 1;\n      }\n      return 0;\n    }\n  }]);\n  return Hero;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/types/hero.js?");
-
-/***/ }),
-
-/***/ "./src/avo/entity/types/shot.js":
-/*!**************************************!*\
-  !*** ./src/avo/entity/types/shot.js ***!
-  \**************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Shot)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar Shot = /*#__PURE__*/function (_Entity) {\n  _inherits(Shot, _Entity);\n  var _super = _createSuper(Shot);\n  function Shot(app) {\n    var _this;\n    var x = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var y = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var rotation = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 0;\n    var source = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : undefined;\n    _classCallCheck(this, Shot);\n    _this = _super.call(this, app);\n    _this._type = 'shot';\n    _this.source = source; // The entity that this shot came from\n\n    _this.colour = '#cc4';\n    _this.solid = false;\n    _this.movable = true;\n    _this.size = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE / 2;\n    _this.x = x;\n    _this.y = y;\n    _this.rotation = rotation;\n    _this.moveAcceleration = 1;\n    _this.moveDeceleration = 0;\n    _this.moveMaxSpeed = _this.size * 2;\n    _this.distance = _this.size * 1;\n    return _this;\n  }\n  _createClass(Shot, [{\n    key: \"onCollision\",\n    value: function onCollision(target, collisionCorrection) {\n      _get(_getPrototypeOf(Shot.prototype), \"onCollision\", this).call(this, target, collisionCorrection);\n      if (this.source !== target && target.solid) {\n        target.applyEffect({\n          name: 'damage'\n        }, this);\n        this._expired = true;\n      }\n    }\n  }, {\n    key: \"play\",\n    value: function play() {\n      _get(_getPrototypeOf(Shot.prototype), \"play\", this).call(this);\n      var moveAcceleration = this.moveAcceleration || 0;\n      this.moveX += moveAcceleration * Math.cos(this.rotation);\n      this.moveY += moveAcceleration * Math.sin(this.rotation);\n      this.distance -= moveAcceleration;\n      if (this.distance <= 0) this._expired = true;\n    }\n  }]);\n  return Shot;\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/types/shot.js?");
-
-/***/ }),
-
-/***/ "./src/avo/entity/types/wall.js":
-/*!**************************************!*\
-  !*** ./src/avo/entity/types/wall.js ***!
-  \**************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Wall)\n/* harmony export */ });\n/* harmony import */ var _avo_entity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/entity */ \"./src/avo/entity/index.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\nvar Wall = /*#__PURE__*/function (_Entity) {\n  _inherits(Wall, _Entity);\n  var _super = _createSuper(Wall);\n  function Wall(app) {\n    var _this;\n    var col = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;\n    var row = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : 0;\n    var width = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : 1;\n    var height = arguments.length > 4 && arguments[4] !== undefined ? arguments[4] : 1;\n    var cutCorner = arguments.length > 5 && arguments[5] !== undefined ? arguments[5] : false;\n    _classCallCheck(this, Wall);\n    _this = _super.call(this, app);\n    _this._type = 'wall';\n    _this.colour = '#888';\n    _this.solid = true;\n    _this.movable = false;\n    _this.x = col * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.y = row * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE;\n    _this.shape = _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.SHAPES.POLYGON;\n    _this.shapePolygonPath = [];\n    if (cutCorner !== 'nw') _this.shapePolygonPath.push(0, 0);\n    if (cutCorner !== 'ne') _this.shapePolygonPath.push(width * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, 0);\n    if (cutCorner !== 'se') _this.shapePolygonPath.push(width * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE, height * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    if (cutCorner !== 'sw') _this.shapePolygonPath.push(0, height * _avo_constants_js__WEBPACK_IMPORTED_MODULE_1__.TILE_SIZE);\n    return _this;\n  }\n  return _createClass(Wall);\n}(_avo_entity__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/entity/types/wall.js?");
 
 /***/ }),
 
@@ -170,6 +170,16 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/avo/rule/examples/zelda-controls.js":
+/*!*************************************************!*\
+  !*** ./src/avo/rule/examples/zelda-controls.js ***!
+  \*************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ZeldaControls)\n/* harmony export */ });\n/* harmony import */ var _avo_rule__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/rule */ \"./src/avo/rule/index.js\");\n/* harmony import */ var _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics.js */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n/*\r\nStandard player controls for top-down adventure games.\r\n */\nvar ZeldaControls = /*#__PURE__*/function (_Rule) {\n  _inherits(ZeldaControls, _Rule);\n  var _super = _createSuper(ZeldaControls);\n  function ZeldaControls(app) {\n    var _this;\n    _classCallCheck(this, ZeldaControls);\n    _this = _super.call(this, app);\n    _this._type = 'zelda-controls';\n    _this.inputTap = false;\n    _this.onPointerTap = _this.onPointerTap.bind(_assertThisInitialized(_this));\n    app.addEventListener('pointertap', _this.onPointerTap);\n    return _this;\n  }\n  _createClass(ZeldaControls, [{\n    key: \"deconstructor\",\n    value: function deconstructor() {\n      app.removeEventListener('pointertap', this.onPointerTap);\n    }\n  }, {\n    key: \"play\",\n    value: function play() {\n      var app = this._app;\n      var hero = app.hero;\n      _get(_getPrototypeOf(ZeldaControls.prototype), \"play\", this).call(this);\n      if (hero) {\n        var _app$playerInput = app.playerInput,\n          keysPressed = _app$playerInput.keysPressed,\n          pointerCurrent = _app$playerInput.pointerCurrent,\n          pointerStart = _app$playerInput.pointerStart,\n          pointerState = _app$playerInput.pointerState;\n        var intent = undefined;\n        var directionX = 0;\n        var directionY = 0;\n        if (pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.POINTER_STATES.POINTER_DOWN) {\n          // Get pointer input if there's any.\n\n          var distX = pointerCurrent.x - pointerStart.x;\n          var distY = pointerCurrent.y - pointerStart.y;\n          var pointerDistance = Math.sqrt(distX * distX + distY * distY);\n          // const movementAngle = Math.atan2(distY, distX)\n\n          if (pointerDistance > _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.POINTER_DEADZONE_RADIUS) {\n            directionX = distX / pointerDistance;\n            directionY = distY / pointerDistance;\n          }\n        } else {\n          // Otherwise, check for keyboard input.\n\n          if (keysPressed['ArrowRight']) directionX++;\n          if (keysPressed['ArrowDown']) directionY++;\n          if (keysPressed['ArrowLeft']) directionX--;\n          if (keysPressed['ArrowUp']) directionY--;\n        }\n        if (keysPressed['z'] && !keysPressed['z'].acknowledged || keysPressed['Z'] && !keysPressed['Z'].acknowledged || this.inputTap) {\n          intent = {\n            name: 'dash',\n            directionX: directionX,\n            directionY: directionY\n          };\n          if (keysPressed['z']) keysPressed['z'].acknowledged = true;\n          if (keysPressed['Z']) keysPressed['Z'].acknowledged = true;\n          this.inputTap = false;\n        } else if (directionX || directionY) {\n          intent = {\n            name: 'move',\n            directionX: directionX,\n            directionY: directionY\n          };\n        }\n        hero.intent = intent;\n      }\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var hero = this._app.hero;\n      if (layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.LAYERS.OVERLAY) {\n        this.paintUIData();\n        this.paintPointerInput();\n      } else if (layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.LAYERS.BOTTOM) {\n        this.paintLineOfSight(hero);\n      }\n    }\n\n    /*\r\n    Draw UI data, such as Hero health.\r\n     */\n  }, {\n    key: \"paintUIData\",\n    value: function paintUIData() {\n      var _hero$action;\n      var c2d = this._app.canvas2d;\n      var hero = this._app.hero;\n      var X_OFFSET = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 1.5;\n      var Y_OFFSET = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * -1.0;\n      var LEFT = X_OFFSET;\n      var RIGHT = this._app.canvasWidth - X_OFFSET;\n      var BOTTOM = this._app.canvasHeight + Y_OFFSET;\n      c2d.font = '2em Source Code Pro';\n      c2d.textBaseline = 'bottom';\n      c2d.lineWidth = 8;\n      var health = Math.max(hero === null || hero === void 0 ? void 0 : hero.health, 0) || 0;\n      var text = '❤️'.repeat(health);\n      c2d.textAlign = 'left';\n      c2d.strokeStyle = '#fff';\n      c2d.strokeText(text, LEFT, BOTTOM);\n      c2d.fillStyle = '#c44';\n      c2d.fillText(text, LEFT, BOTTOM);\n      text = (hero === null || hero === void 0 || (_hero$action = hero.action) === null || _hero$action === void 0 ? void 0 : _hero$action.name) + ' (' + (hero === null || hero === void 0 ? void 0 : hero.moveSpeed.toFixed(2)) + ')';\n      c2d.textAlign = 'right';\n      c2d.strokeStyle = '#fff';\n      c2d.strokeText(text, RIGHT, BOTTOM);\n      c2d.fillStyle = '#c44';\n      c2d.fillText(text, RIGHT, BOTTOM);\n    }\n\n    /*\r\n    Draw pointer input, if any. This helps players get visual feedback on their\r\n    touchscreens.\r\n     */\n  }, {\n    key: \"paintPointerInput\",\n    value: function paintPointerInput() {\n      var c2d = this._app.canvas2d;\n      var _this$_app$playerInpu = this._app.playerInput,\n        pointerCurrent = _this$_app$playerInpu.pointerCurrent,\n        pointerStart = _this$_app$playerInpu.pointerStart,\n        pointerState = _this$_app$playerInpu.pointerState;\n      var START_POINT_RADIUS = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 1,\n        CURRENT_POINT_RADIUS = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 0.5;\n      if (pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.POINTER_STATES.POINTER_DOWN) {\n        c2d.lineWidth = Math.floor(Math.min(_avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 0.125, 2));\n        c2d.fillStyle = '#80808080';\n        c2d.strokeStyle = '#80808080';\n        c2d.beginPath();\n        c2d.arc(pointerStart.x, pointerStart.y, START_POINT_RADIUS, 0, 2 * Math.PI);\n        c2d.stroke();\n        c2d.beginPath();\n        c2d.arc(pointerCurrent.x, pointerCurrent.y, CURRENT_POINT_RADIUS, 0, 2 * Math.PI);\n        c2d.fill();\n        c2d.beginPath();\n        c2d.moveTo(pointerStart.x, pointerStart.y);\n        c2d.lineTo(pointerCurrent.x, pointerCurrent.y);\n        c2d.stroke();\n      }\n    }\n\n    /*\r\n    Draw a line of sight (cast a ray) starting from a specified Entity (usually the\r\n    hero), in the direction they're facing.\r\n     */\n  }, {\n    key: \"paintLineOfSight\",\n    value: function paintLineOfSight(srcEntity) {\n      if (!srcEntity) return;\n      var c2d = this._app.canvas2d;\n      var camera = this._app.camera;\n      var entities = this._app.entities;\n      var MAX_LINE_OF_SIGHT_DISTANCE = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 5;\n      this._app.applyCameraTransforms();\n\n      // Intended line of sight, i.e. a ray starting from the hero/source Entity.\n      var lineOfSight = {\n        start: {\n          x: srcEntity.x,\n          y: srcEntity.y\n        },\n        end: {\n          x: srcEntity.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(srcEntity.rotation),\n          y: srcEntity.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(srcEntity.rotation)\n        }\n      };\n      var actualLineOfSightEndPoint = undefined;\n\n      // For each other Entity, see if it intersects with the source Entity's LOS\n      entities.forEach(function (entity) {\n        if (entity === srcEntity) return;\n\n        // TODO: check for opaqueness and/or if the entity is visible.\n\n        var vertices = entity.vertices;\n        if (vertices.length < 2) return;\n\n        // Every entity has a \"shape\" that can be represented by a polygon.\n        // (Yes, even circles.) Check each segment (aka edge aka side) of the\n        // polygon.\n        for (var i = 0; i < vertices.length; i++) {\n          var segment = {\n            start: {\n              x: vertices[i].x,\n              y: vertices[i].y\n            },\n            end: {\n              x: vertices[(i + 1) % vertices.length].x,\n              y: vertices[(i + 1) % vertices.length].y\n            }\n          };\n\n          // Find the intersection. We want to find the intersection point\n          // closest to the source Entity (the LOS ray's starting point).\n          var intersection = _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].getLineIntersection(lineOfSight, segment);\n          if (!actualLineOfSightEndPoint || intersection && intersection.distanceFactor < actualLineOfSightEndPoint.distanceFactor) {\n            actualLineOfSightEndPoint = intersection;\n          }\n        }\n      });\n      if (!actualLineOfSightEndPoint) {\n        actualLineOfSightEndPoint = {\n          x: srcEntity.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(srcEntity.rotation),\n          y: srcEntity.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(srcEntity.rotation)\n        };\n      }\n\n      // Expected line of sight\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x, lineOfSight.start.y);\n      c2d.lineTo(lineOfSight.end.x, lineOfSight.end.y);\n      c2d.strokeStyle = '#c08080';\n      c2d.lineWidth = 3;\n      c2d.stroke();\n      c2d.setLineDash([]);\n\n      // Actual line of sight\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x, lineOfSight.start.y);\n      c2d.lineTo(actualLineOfSightEndPoint.x, actualLineOfSightEndPoint.y);\n      c2d.strokeStyle = '#3399ff';\n      c2d.lineWidth = 3;\n      c2d.stroke();\n\n      // Expected end of line of sight\n      c2d.beginPath();\n      c2d.arc(lineOfSight.end.x, lineOfSight.end.y, 4, 0, 2 * Math.PI);\n      c2d.fillStyle = '#c08080';\n      c2d.fill();\n\n      // Actual end of line of sight\n      c2d.beginPath();\n      c2d.arc(actualLineOfSightEndPoint.x, actualLineOfSightEndPoint.y, 8, 0, 2 * Math.PI);\n      c2d.fillStyle = '#3399ff';\n      c2d.fill();\n      this._app.undoCameraTransforms();\n    }\n  }, {\n    key: \"onPointerTap\",\n    value: function onPointerTap() {\n      this.inputTap = true;\n    }\n  }]);\n  return ZeldaControls;\n}(_avo_rule__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/rule/examples/zelda-controls.js?");
+
+/***/ }),
+
 /***/ "./src/avo/rule/index.js":
 /*!*******************************!*\
   !*** ./src/avo/rule/index.js ***!
@@ -190,13 +200,13 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/avo/rule/types/zelda-controls.js":
-/*!**********************************************!*\
-  !*** ./src/avo/rule/types/zelda-controls.js ***!
-  \**********************************************/
+/***/ "./src/avo/story/examples/example-story.js":
+/*!*************************************************!*\
+  !*** ./src/avo/story/examples/example-story.js ***!
+  \*************************************************/
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ZeldaControls)\n/* harmony export */ });\n/* harmony import */ var _avo_rule__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! @avo/rule */ \"./src/avo/rule/index.js\");\n/* harmony import */ var _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/physics.js */ \"./src/avo/physics.js\");\n/* harmony import */ var _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/constants.js */ \"./src/avo/constants.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n/*\r\nStandard player controls for top-down adventure games.\r\n */\nvar ZeldaControls = /*#__PURE__*/function (_Rule) {\n  _inherits(ZeldaControls, _Rule);\n  var _super = _createSuper(ZeldaControls);\n  function ZeldaControls(app) {\n    var _this;\n    _classCallCheck(this, ZeldaControls);\n    _this = _super.call(this, app);\n    _this._type = 'zelda-controls';\n    _this.inputTap = false;\n    _this.onPointerTap = _this.onPointerTap.bind(_assertThisInitialized(_this));\n    app.addEventListener('pointertap', _this.onPointerTap);\n    return _this;\n  }\n  _createClass(ZeldaControls, [{\n    key: \"deconstructor\",\n    value: function deconstructor() {\n      app.removeEventListener('pointertap', this.onPointerTap);\n    }\n  }, {\n    key: \"play\",\n    value: function play() {\n      var app = this._app;\n      var hero = app.hero;\n      _get(_getPrototypeOf(ZeldaControls.prototype), \"play\", this).call(this);\n      if (hero) {\n        var _app$playerInput = app.playerInput,\n          keysPressed = _app$playerInput.keysPressed,\n          pointerCurrent = _app$playerInput.pointerCurrent,\n          pointerStart = _app$playerInput.pointerStart,\n          pointerState = _app$playerInput.pointerState;\n        var intent = undefined;\n        var directionX = 0;\n        var directionY = 0;\n        if (pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.POINTER_STATES.POINTER_DOWN) {\n          // Get pointer input if there's any.\n\n          var distX = pointerCurrent.x - pointerStart.x;\n          var distY = pointerCurrent.y - pointerStart.y;\n          var pointerDistance = Math.sqrt(distX * distX + distY * distY);\n          // const movementAngle = Math.atan2(distY, distX)\n\n          if (pointerDistance > _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.POINTER_DEADZONE_RADIUS) {\n            directionX = distX / pointerDistance;\n            directionY = distY / pointerDistance;\n          }\n        } else {\n          // Otherwise, check for keyboard input.\n\n          if (keysPressed['ArrowRight']) directionX++;\n          if (keysPressed['ArrowDown']) directionY++;\n          if (keysPressed['ArrowLeft']) directionX--;\n          if (keysPressed['ArrowUp']) directionY--;\n        }\n        if (keysPressed['z'] && !keysPressed['z'].acknowledged || keysPressed['Z'] && !keysPressed['Z'].acknowledged || this.inputTap) {\n          intent = {\n            name: 'dash',\n            directionX: directionX,\n            directionY: directionY\n          };\n          if (keysPressed['z']) keysPressed['z'].acknowledged = true;\n          if (keysPressed['Z']) keysPressed['Z'].acknowledged = true;\n          this.inputTap = false;\n        } else if (directionX || directionY) {\n          intent = {\n            name: 'move',\n            directionX: directionX,\n            directionY: directionY\n          };\n        }\n        hero.intent = intent;\n      }\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var layer = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;\n      var hero = this._app.hero;\n      if (layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.LAYERS.OVERLAY) {\n        this.paintUIData();\n        this.paintPointerInput();\n      } else if (layer === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.LAYERS.BOTTOM) {\n        this.paintLineOfSight(hero);\n      }\n    }\n\n    /*\r\n    Draw UI data, such as Hero health.\r\n     */\n  }, {\n    key: \"paintUIData\",\n    value: function paintUIData() {\n      var _hero$action;\n      var c2d = this._app.canvas2d;\n      var hero = this._app.hero;\n      var X_OFFSET = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 1.5;\n      var Y_OFFSET = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * -1.0;\n      var LEFT = X_OFFSET;\n      var RIGHT = this._app.canvasWidth - X_OFFSET;\n      var BOTTOM = this._app.canvasHeight + Y_OFFSET;\n      c2d.font = '2em Source Code Pro';\n      c2d.textBaseline = 'bottom';\n      c2d.lineWidth = 8;\n      var health = Math.max(hero === null || hero === void 0 ? void 0 : hero.health, 0) || 0;\n      var text = '❤️'.repeat(health);\n      c2d.textAlign = 'left';\n      c2d.strokeStyle = '#fff';\n      c2d.strokeText(text, LEFT, BOTTOM);\n      c2d.fillStyle = '#c44';\n      c2d.fillText(text, LEFT, BOTTOM);\n      text = (hero === null || hero === void 0 || (_hero$action = hero.action) === null || _hero$action === void 0 ? void 0 : _hero$action.name) + ' (' + (hero === null || hero === void 0 ? void 0 : hero.moveSpeed.toFixed(2)) + ')';\n      c2d.textAlign = 'right';\n      c2d.strokeStyle = '#fff';\n      c2d.strokeText(text, RIGHT, BOTTOM);\n      c2d.fillStyle = '#c44';\n      c2d.fillText(text, RIGHT, BOTTOM);\n    }\n\n    /*\r\n    Draw pointer input, if any. This helps players get visual feedback on their\r\n    touchscreens.\r\n     */\n  }, {\n    key: \"paintPointerInput\",\n    value: function paintPointerInput() {\n      var c2d = this._app.canvas2d;\n      var _this$_app$playerInpu = this._app.playerInput,\n        pointerCurrent = _this$_app$playerInpu.pointerCurrent,\n        pointerStart = _this$_app$playerInpu.pointerStart,\n        pointerState = _this$_app$playerInpu.pointerState;\n      var START_POINT_RADIUS = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 1,\n        CURRENT_POINT_RADIUS = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 0.5;\n      if (pointerState === _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.POINTER_STATES.POINTER_DOWN) {\n        c2d.lineWidth = Math.floor(Math.min(_avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 0.125, 2));\n        c2d.fillStyle = '#80808080';\n        c2d.strokeStyle = '#80808080';\n        c2d.beginPath();\n        c2d.arc(pointerStart.x, pointerStart.y, START_POINT_RADIUS, 0, 2 * Math.PI);\n        c2d.stroke();\n        c2d.beginPath();\n        c2d.arc(pointerCurrent.x, pointerCurrent.y, CURRENT_POINT_RADIUS, 0, 2 * Math.PI);\n        c2d.fill();\n        c2d.beginPath();\n        c2d.moveTo(pointerStart.x, pointerStart.y);\n        c2d.lineTo(pointerCurrent.x, pointerCurrent.y);\n        c2d.stroke();\n      }\n    }\n\n    /*\r\n    Draw a line of sight (cast a ray) starting from a specified Entity (usually the\r\n    hero), in the direction they're facing.\r\n     */\n  }, {\n    key: \"paintLineOfSight\",\n    value: function paintLineOfSight(srcEntity) {\n      if (!srcEntity) return;\n      var c2d = this._app.canvas2d;\n      var camera = this._app.camera;\n      var entities = this._app.entities;\n      var MAX_LINE_OF_SIGHT_DISTANCE = _avo_constants_js__WEBPACK_IMPORTED_MODULE_2__.TILE_SIZE * 5;\n      this._app.applyCameraTransforms();\n\n      // Intended line of sight, i.e. a ray starting from the hero/source Entity.\n      var lineOfSight = {\n        start: {\n          x: srcEntity.x,\n          y: srcEntity.y\n        },\n        end: {\n          x: srcEntity.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(srcEntity.rotation),\n          y: srcEntity.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(srcEntity.rotation)\n        }\n      };\n      var actualLineOfSightEndPoint = undefined;\n\n      // For each other Entity, see if it intersects with the source Entity's LOS\n      entities.forEach(function (entity) {\n        if (entity === srcEntity) return;\n\n        // TODO: check for opaqueness and/or if the entity is visible.\n\n        var vertices = entity.vertices;\n        if (vertices.length < 2) return;\n\n        // Every entity has a \"shape\" that can be represented by a polygon.\n        // (Yes, even circles.) Check each segment (aka edge aka side) of the\n        // polygon.\n        for (var i = 0; i < vertices.length; i++) {\n          var segment = {\n            start: {\n              x: vertices[i].x,\n              y: vertices[i].y\n            },\n            end: {\n              x: vertices[(i + 1) % vertices.length].x,\n              y: vertices[(i + 1) % vertices.length].y\n            }\n          };\n\n          // Find the intersection. We want to find the intersection point\n          // closest to the source Entity (the LOS ray's starting point).\n          var intersection = _avo_physics_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].getLineIntersection(lineOfSight, segment);\n          if (!actualLineOfSightEndPoint || intersection && intersection.distanceFactor < actualLineOfSightEndPoint.distanceFactor) {\n            actualLineOfSightEndPoint = intersection;\n          }\n        }\n      });\n      if (!actualLineOfSightEndPoint) {\n        actualLineOfSightEndPoint = {\n          x: srcEntity.x + MAX_LINE_OF_SIGHT_DISTANCE * Math.cos(srcEntity.rotation),\n          y: srcEntity.y + MAX_LINE_OF_SIGHT_DISTANCE * Math.sin(srcEntity.rotation)\n        };\n      }\n\n      // Expected line of sight\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x, lineOfSight.start.y);\n      c2d.lineTo(lineOfSight.end.x, lineOfSight.end.y);\n      c2d.strokeStyle = '#c08080';\n      c2d.lineWidth = 3;\n      c2d.stroke();\n      c2d.setLineDash([]);\n\n      // Actual line of sight\n      c2d.beginPath();\n      c2d.moveTo(lineOfSight.start.x, lineOfSight.start.y);\n      c2d.lineTo(actualLineOfSightEndPoint.x, actualLineOfSightEndPoint.y);\n      c2d.strokeStyle = '#3399ff';\n      c2d.lineWidth = 3;\n      c2d.stroke();\n\n      // Expected end of line of sight\n      c2d.beginPath();\n      c2d.arc(lineOfSight.end.x, lineOfSight.end.y, 4, 0, 2 * Math.PI);\n      c2d.fillStyle = '#c08080';\n      c2d.fill();\n\n      // Actual end of line of sight\n      c2d.beginPath();\n      c2d.arc(actualLineOfSightEndPoint.x, actualLineOfSightEndPoint.y, 8, 0, 2 * Math.PI);\n      c2d.fillStyle = '#3399ff';\n      c2d.fill();\n      this._app.undoCameraTransforms();\n    }\n  }, {\n    key: \"onPointerTap\",\n    value: function onPointerTap() {\n      this.inputTap = true;\n    }\n  }]);\n  return ZeldaControls;\n}(_avo_rule__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/rule/types/zelda-controls.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ExampleStory)\n/* harmony export */ });\n/* harmony import */ var _story__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../story */ \"./src/avo/story/story.js\");\n/* harmony import */ var _avo_image_asset_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/image-asset.js */ \"./src/avo/image-asset.js\");\n/* harmony import */ var _avo_entity_examples_hero_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/entity/examples/hero.js */ \"./src/avo/entity/examples/hero.js\");\n/* harmony import */ var _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/entity/examples/wall.js */ \"./src/avo/entity/examples/wall.js\");\n/* harmony import */ var _avo_entity_examples_ball_js__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/entity/examples/ball.js */ \"./src/avo/entity/examples/ball.js\");\n/* harmony import */ var _avo_entity_examples_enemy_js__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/entity/examples/enemy.js */ \"./src/avo/entity/examples/enemy.js\");\n/* harmony import */ var _avo_rule_examples_zelda_controls_js__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! @avo/rule/examples/zelda-controls.js */ \"./src/avo/rule/examples/zelda-controls.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n\n\n\nvar ExampleStory = /*#__PURE__*/function (_Story) {\n  _inherits(ExampleStory, _Story);\n  var _super = _createSuper(ExampleStory);\n  function ExampleStory(app) {\n    var _this;\n    _classCallCheck(this, ExampleStory);\n    _this = _super.call(this, app);\n    _this.assets = {\n      \"hero-4dir\": new _avo_image_asset_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]('assets/avo-sprites-2022-05-samiel.png')\n    };\n    return _this;\n  }\n  _createClass(ExampleStory, [{\n    key: \"start\",\n    value: function start() {\n      _get(_getPrototypeOf(ExampleStory.prototype), \"start\", this).call(this);\n      this.generate_zelda_default();\n    }\n\n    /*\r\n    Default top-down adventure level.\r\n     */\n  }, {\n    key: \"generate_zelda_default\",\n    value: function generate_zelda_default() {\n      var app = this._app;\n      app.hero = app.addEntity(new _avo_entity_examples_hero_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"](app, 11, 1));\n      app.camera.target = app.hero;\n      app.addRule(new _avo_rule_examples_zelda_controls_js__WEBPACK_IMPORTED_MODULE_6__[\"default\"](app));\n      app.addEntity(new _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 0, 0, 1, 23)); // West Wall\n      app.addEntity(new _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 22, 0, 1, 23)); // East Wall\n      app.addEntity(new _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 1, 0, 21, 1)); // North Wall\n      app.addEntity(new _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 1, 22, 21, 1)); // South Wall\n      app.addEntity(new _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 3, 2, 3, 1));\n      app.addEntity(new _avo_entity_examples_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 3, 4, 3, 1));\n      app.addEntity(new _avo_entity_examples_ball_js__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 8, 6));\n      var enemy = app.addEntity(new _avo_entity_examples_enemy_js__WEBPACK_IMPORTED_MODULE_5__[\"default\"](app, 4, 8));\n      enemy.rotation = -45 / 180 * Math.PI;\n    }\n  }]);\n  return ExampleStory;\n}(_story__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/examples/example-story.js?");
 
 /***/ }),
 
@@ -217,16 +227,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ Story)\n/* harmony export */ });\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nvar Story = /*#__PURE__*/function () {\n  function Story(app) {\n    _classCallCheck(this, Story);\n    this._app = app;\n    this.assets = {};\n  }\n\n  /*\r\n  Section: Main Scripts\r\n  ----------------------------------------------------------------------------\r\n   */\n  _createClass(Story, [{\n    key: \"start\",\n    value: function start() {\n      this.reset();\n    }\n  }, {\n    key: \"reset\",\n    value: function reset() {\n      var app = this._app;\n      app.hero = undefined;\n      app.clearEntities();\n      app.resetGameMap();\n      app.clearRules();\n      app.resetCamera();\n      app.resetPlayerInput();\n      app.setInteractionMenu(false);\n    }\n  }, {\n    key: \"reload\",\n    value: function reload() {\n      this.start();\n    }\n  }]);\n  return Story;\n}();\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/story.js?");
-
-/***/ }),
-
-/***/ "./src/avo/story/types/example-story.js":
-/*!**********************************************!*\
-  !*** ./src/avo/story/types/example-story.js ***!
-  \**********************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (/* binding */ ExampleStory)\n/* harmony export */ });\n/* harmony import */ var _story__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../story */ \"./src/avo/story/story.js\");\n/* harmony import */ var _avo_image_asset_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @avo/image-asset.js */ \"./src/avo/image-asset.js\");\n/* harmony import */ var _avo_entity_types_hero_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! @avo/entity/types/hero.js */ \"./src/avo/entity/types/hero.js\");\n/* harmony import */ var _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @avo/entity/types/wall.js */ \"./src/avo/entity/types/wall.js\");\n/* harmony import */ var _avo_entity_types_ball_js__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! @avo/entity/types/ball.js */ \"./src/avo/entity/types/ball.js\");\n/* harmony import */ var _avo_entity_types_enemy_js__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @avo/entity/types/enemy.js */ \"./src/avo/entity/types/enemy.js\");\n/* harmony import */ var _avo_rule_types_zelda_controls_js__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! @avo/rule/types/zelda-controls.js */ \"./src/avo/rule/types/zelda-controls.js\");\nfunction _typeof(o) { \"@babel/helpers - typeof\"; return _typeof = \"function\" == typeof Symbol && \"symbol\" == typeof Symbol.iterator ? function (o) { return typeof o; } : function (o) { return o && \"function\" == typeof Symbol && o.constructor === Symbol && o !== Symbol.prototype ? \"symbol\" : typeof o; }, _typeof(o); }\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, _toPropertyKey(descriptor.key), descriptor); } }\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); Object.defineProperty(Constructor, \"prototype\", { writable: false }); return Constructor; }\nfunction _toPropertyKey(t) { var i = _toPrimitive(t, \"string\"); return \"symbol\" == _typeof(i) ? i : String(i); }\nfunction _toPrimitive(t, r) { if (\"object\" != _typeof(t) || !t) return t; var e = t[Symbol.toPrimitive]; if (void 0 !== e) { var i = e.call(t, r || \"default\"); if (\"object\" != _typeof(i)) return i; throw new TypeError(\"@@toPrimitive must return a primitive value.\"); } return (\"string\" === r ? String : Number)(t); }\nfunction _get() { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get.bind(); } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(arguments.length < 3 ? target : receiver); } return desc.value; }; } return _get.apply(this, arguments); }\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); Object.defineProperty(subClass, \"prototype\", { writable: false }); if (superClass) _setPrototypeOf(subClass, superClass); }\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf ? Object.setPrototypeOf.bind() : function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\nfunction _createSuper(Derived) { var hasNativeReflectConstruct = _isNativeReflectConstruct(); return function _createSuperInternal() { var Super = _getPrototypeOf(Derived), result; if (hasNativeReflectConstruct) { var NewTarget = _getPrototypeOf(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return _possibleConstructorReturn(this, result); }; }\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } else if (call !== void 0) { throw new TypeError(\"Derived constructors may only return object or undefined\"); } return _assertThisInitialized(self); }\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\nfunction _isNativeReflectConstruct() { if (typeof Reflect === \"undefined\" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === \"function\") return true; try { Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function () {})); return true; } catch (e) { return false; } }\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf.bind() : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\n\n\n\n\n\n\nvar ExampleStory = /*#__PURE__*/function (_Story) {\n  _inherits(ExampleStory, _Story);\n  var _super = _createSuper(ExampleStory);\n  function ExampleStory(app) {\n    var _this;\n    _classCallCheck(this, ExampleStory);\n    _this = _super.call(this, app);\n    _this.assets = {\n      \"hero-4dir\": new _avo_image_asset_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]('assets/avo-sprites-2022-05-samiel.png')\n    };\n    return _this;\n  }\n  _createClass(ExampleStory, [{\n    key: \"start\",\n    value: function start() {\n      _get(_getPrototypeOf(ExampleStory.prototype), \"start\", this).call(this);\n      this.generate_zelda_default();\n    }\n\n    /*\r\n    Default top-down adventure level.\r\n     */\n  }, {\n    key: \"generate_zelda_default\",\n    value: function generate_zelda_default() {\n      var app = this._app;\n      app.hero = app.addEntity(new _avo_entity_types_hero_js__WEBPACK_IMPORTED_MODULE_2__[\"default\"](app, 11, 1));\n      app.camera.target = app.hero;\n      app.addRule(new _avo_rule_types_zelda_controls_js__WEBPACK_IMPORTED_MODULE_6__[\"default\"](app));\n      app.addEntity(new _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 0, 0, 1, 23)); // West Wall\n      app.addEntity(new _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 22, 0, 1, 23)); // East Wall\n      app.addEntity(new _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 1, 0, 21, 1)); // North Wall\n      app.addEntity(new _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 1, 22, 21, 1)); // South Wall\n      app.addEntity(new _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 3, 2, 3, 1));\n      app.addEntity(new _avo_entity_types_wall_js__WEBPACK_IMPORTED_MODULE_3__[\"default\"](app, 3, 4, 3, 1));\n      app.addEntity(new _avo_entity_types_ball_js__WEBPACK_IMPORTED_MODULE_4__[\"default\"](app, 8, 6));\n      var enemy = app.addEntity(new _avo_entity_types_enemy_js__WEBPACK_IMPORTED_MODULE_5__[\"default\"](app, 4, 8));\n      enemy.rotation = -45 / 180 * Math.PI;\n    }\n  }]);\n  return ExampleStory;\n}(_story__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n\n//# sourceURL=webpack://avo-adventure-mk3/./src/avo/story/types/example-story.js?");
 
 /***/ }),
