+++ src/avo/entity/examples/hero.js
