+++ src/avo/entity/examples/shot.js
