+++ src/avo/rule/examples/zelda-controls.js
